% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{addConfboundsToVar}
\alias{addConfboundsToVar}
\title{Format confidence bounds for a variable into string format bracketed notation}
\usage{
addConfboundsToVar(var, confLower, confUpper, sigdig = 3)
}
\arguments{
\item{var}{a list of values for the varaible}

\item{confLower}{the lower bounds for each of the values}

\item{confUpper}{the upper bounds for each of the values}

\item{sigdig}{the number of significant digits}
}
\description{
Format confidence bounds for a variable into string format bracketed notation
}
