% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovar.R
\name{addCovariate}
\alias{addCovariate}
\title{Add covariate expression to a function string}
\usage{
addCovariate(funstring, varName, covariate, theta, isLog)
}
\arguments{
\item{funstring}{a string giving the expression that needs to be modified}

\item{varName}{the variable to which the given string corresponds to in the model expression}

\item{covariate}{the covariate expression that needs to be added (at the appropriate place)}

\item{theta}{a list of names of the 'theta' parameters in the 'fit' object}

\item{isLog}{a boolean signifying the presence of log-transformation in the funstring}
}
\value{
returns the modified string with the covariate added to function string
}
\description{
Add covariate expression to a function string
}
\author{
Vipul Mann, Matthew Fidler
}
