% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovar.R
\name{addCovMultiple}
\alias{addCovMultiple}
\title{Add multiple covariates to a given model, sequentially or all at once}
\usage{
addCovMultiple(covInfo, fitobject, indep = TRUE)
}
\arguments{
\item{covInfo}{a list containing information about each variable-covariate pair}

\item{fitobject}{an nlmixr 'fit' object}

\item{indep}{a boolean indicating if the covariates should be added independently, or sequentially (append to the previous model); default is TRUE}
}
\description{
Add multiple covariates to a given model, sequentially or all at once
}
\author{
Vipul Mann, Matthew Fidler
}
