% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{as.data.frame.focei.fit}
\alias{as.data.frame.focei.fit}
\title{Convert focei fit to a data.frame}
\usage{
\method{as.data.frame}{focei.fit}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{focei fit object}

\item{row.names}{row names for the data frame.}

\item{optional}{If TRUE setting row names and column names is
optional.}

\item{...}{Other parameters}
}
\value{
A data frame with no information about the focei fit
}
\description{
Convert focei fit to a data.frame
}
\author{
Matthew L. Fidler
}
