% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynmodel.R
\name{as.dynmodel}
\alias{as.dynmodel}
\title{Convert fit to classic dynmodel object}
\usage{
as.dynmodel(x)
}
\arguments{
\item{x}{nlmixr object to convert to dynmodel object}
}
\value{
dynmodel
}
\description{
Convert fit to classic dynmodel object
}
\author{
Matthew Fidler
}
