% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.focei.dynmodel}
\alias{as.focei.dynmodel}
\title{Output nlmixr format for dynmodel}
\usage{
\method{as.focei}{dynmodel}(.dynmodelObject, .nlmixrObject, .data, .time,
  .theta, .fit, .message, .inits.err, .cov, .sgy, .dynmodelControl,
  .nobs2 = 0, .pt = proc.time(), .rxControl = RxODE::rxControl())
}
\arguments{
\item{.dynmodelObject}{return object from}
}
\description{
Output nlmixr format for dynmodel
}
