% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R, R/nlme_fit.R
\name{as.focei}
\alias{as.focei}
\alias{as.focei.nlmixr_nlme}
\title{Convert fit to FOCEi style fit}
\usage{
as.focei(object, uif, pt = proc.time(), ...)

\method{as.focei}{nlmixr_nlme}(object, uif, pt = proc.time(), ...)
}
\arguments{
\item{object}{Fit object to convert to FOCEi-style fit.}

\item{uif}{Unified Interface Function}

\item{pt}{Proc time object}

\item{...}{Other Parameters}
}
\value{
A FOCEi fit style object.
}
\description{
Convert fit to FOCEi style fit
}
\author{
Matthew L. Fidler
}
