% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{as.nlmixrBounds}
\alias{as.nlmixrBounds}
\title{Verify the accuracy of a nlmixrBounds object and update initial conditions,
as required.}
\usage{
as.nlmixrBounds(df)
}
\description{
Verify the accuracy of a nlmixrBounds object and update initial conditions,
as required.
}
