% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable.R
\name{asHux.nlmixrFitCore}
\alias{asHux.nlmixrFitCore}
\title{Change a nlmixr fit object to a huxtable}
\usage{
asHux.nlmixrFitCore(
  ...,
  error_format = "({std.error})",
  error_style = c("stderr", "ci", "statistic", "pvalue"),
  error_pos = c("below", "same", "right"),
  number_format = "\%.3f",
  align = ".",
  pad_decimal = ".",
  ci_level = NULL,
  tidy_args = NULL,
  stars = c(`***` = 0.001, `**` = 0.01, `*` = 0.05),
  bold_signif = NULL,
  borders = 0.4,
  outer_borders = 0.8,
  note = if (is.null(stars)) NULL else "{stars}.",
  statistics = c(N = "nobs", `Objective Function` = "OBJF", "logLik", "AIC"),
  coefs = NULL,
  omit_coefs = NULL,
  na_omit = c("all", "any", "none")
)
}
\arguments{
\item{na_omit}{How NAs are handled in converting to huxtable}
}
\description{
This is a thin layer that differs from
\code{\link[huxtable]{huxreg}} to make it easier to produce
reasonable huxtables with nlmixr fit objects.
}
\details{
\itemize{

\item Drops \code{NA} values from tables, so (\code{NA}) is not
shown for estimates without a standard error.

\item Filters out any blank rows.

\item Removes R.squared from the statistics and replaces with the
objective function value.

\item Adjust broom separators for an easier to read format.

}
}
