% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{assignToEnv}
\alias{assignToEnv}
\title{Assign a set of variables to the nlmixr fit environment}
\usage{
assignToEnv(namedVars, fitobject)
}
\arguments{
\item{namedVars}{a named list of variables that need to be assigned to the given environment}

\item{fitobject}{the nlmixr fit object that contains its environment information}
}
\description{
Assign a set of variables to the nlmixr fit environment
}
