% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bootplot}
\alias{bootplot}
\title{Produce trace-plot for fit if applicable}
\usage{
bootplot(obj, ...)
}
\arguments{
\item{...}{other parameters}

\item{x}{fit object}
}
\value{
Fit traceplot or nothing.
}
\description{
Produce trace-plot for fit if applicable
}
\author{
Vipul Mann,  Matthew L. Fidler
}
