% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapFit}
\alias{bootstrapFit}
\title{Bootstrap nlmixr fit}
\usage{
bootstrapFit(fit, nboot = 500, nSampIndiv, pvalues = NULL, resume = TRUE)
}
\arguments{
\item{fit}{the nlmixr fit object}

\item{nboot}{an integer giving the number of bootstrapped models to be fit; default value is 100}

\item{nSampIndiv}{an integer specifying the number of samples in each bootstrapped sample; default is the number of unique subjects in the original dataset}

\item{pvalues}{a vector of pvalues indicating the probability of each subject to get selected; default value is NULL implying that probability of each subject is the same}

\item{resume}{a boolean that indicates if a previous session has to be resumed; default value is TRUE}
}
\description{
Bootstrap input dataset and rerun the model to get confidence bounds and aggregated parameters
}
\examples{
bootstrapFit(fit)
bootstrapFit(fit, nboot = 5, resume = FALSE) # overwrites any of the existing data or model files
bootstrapFit(fit, nboot = 7) # resumes fitting using the stored data and model files
}
\author{
Vipul Mann, Matthew Fidler
}
