% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collectWarnings}
\alias{collectWarnings}
\title{Collect warnings and just warn once.}
\usage{
collectWarnings(expr)
}
\arguments{
\item{expr}{R expression}
}
\value{
The value of the expression
}
\description{
Collect warnings and just warn once.
}
\author{
Matthew L. Fidler
}
