% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foceiFit.R
\name{constructLinCmt}
\alias{constructLinCmt}
\title{Construct RxODE linCmt function}
\usage{
constructLinCmt(fun)
}
\arguments{
\item{fun}{function to convert to solveC syntax}
}
\value{
SolvedC RxODE object
}
\description{
Construct RxODE linCmt function
}
\author{
Matthew L. Fidler
}
\keyword{internal}
