% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovar.R
\name{covarSearchAuto}
\alias{covarSearchAuto}
\title{Stepwise Covariate Model-selection (SCM) method}
\usage{
covarSearchAuto(
  fit,
  varsVec,
  covarsVec,
  pVal = list(fwd = 0.05, bck = 0.01),
  covInformation = NULL,
  catCovariates = NULL,
  searchType = c("scm", "forward", "backward"),
  restart = FALSE
)
}
\arguments{
\item{fit}{an nlmixr 'fit' object}

\item{varsVec}{a list of candidate variables to which the covariates could be added}

\item{covarsVec}{a list of candidate covariates that need to be tested}

\item{pVal}{a named list with names 'fwd' and 'bck' for specifying the p-values for the forward and backward searches, respectively}

\item{covInformation}{a list containing additionl information on the variables-covariates pairs that should be passed on to addCovMultiple function}

\item{searchType}{one of 'scm', 'forward' and 'backward' to specify the covariate search method; default is 'scm'}

\item{restart}{a boolean that controls if the search should be restarted; default is FALSE}

\item{catCovaraites}{a list of covariates that should be treated as categorical}
}
\description{
Stepwise Covariate Model-selection (SCM) method
}
\examples{
covarSearchAuto(fit, varsVec = c("ka", "cl"), covarsVec = c("WT", "SEX"), catCovariates = c("SEX"))
covarSearchAuto(fit, varsVec = c("ka", "cl"), covarsVec = c("WT", "SEX"), catCovariates = c("SEX"), restart = TRUE)
covarSearchAuto(fit, varsVec = c("ka", "cl"), covarsVec = c("WT", "SEX"), catCovariates = c("SEX"), restart = TRUE, searchType = "forward")
}
\author{
Vipul Mann, Matthew Fidler
}
