% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foceiFit.R
\name{coxBox}
\alias{coxBox}
\title{Cox Box transformation}
\usage{
coxBox(x, lambda = 1)
}
\arguments{
\item{x}{data to transform}

\item{lambda}{Cox-box lambda parameter}
}
\value{
Cox-Box Transformed Data
}
\description{
Cox Box transformation
}
\author{
Matthew L. Fidler
}
