% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem_fit.R
\name{.protectSaem}
\alias{.protectSaem}
\title{SAEM dll prodection from garbage collection}
\usage{
.protectSaem(dll)
}
\arguments{
\item{dll}{dll to protect}
}
\value{
nothing
}
\description{
This protects the saem dll from being prematurely unloaded while
running it due to garbage collection.
}
