% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dynmodel}
\alias{dynmodel}
\title{Fit a non-population dynamic model}
\usage{
dynmodel(system, model, inits, data, fixPars = NULL, control = list(), ...)
}
\arguments{
\item{system}{RxODE object. See \code{\link[RxODE]{RxODE}} for more details.}

\item{model}{Error model.}

\item{inits}{Initial values of system parameters.}

\item{data}{Dataset to estimate. Needs to be RxODE compatible in EVIDs.}

\item{fixPars}{Fixed system parameters. Default is NULL.}

\item{control}{Control options for dynmodel
\code{\link[nlmixr]{dynmodelControl}} .}
}
\description{
Fit a non-population dynamic model
}
\examples{
# dynmodel example --------------------------------------------------------
ode <- "
      kel = CL/V;
      d/dt(X) = -kel*X;
      C=X/V;
      PRED = C
      "
ode_system <- RxODE(model = ode)
model_error_structure <- cp ~ C + add(0.01) + prop(0.01)
inits <- c(CL = 1, V = 10)
control <- dynmodelControl(method = "Nelder-Mead")
fit <-
  dynmodel(
    system = ode_system,
    model = model_error_structure,
    data = Bolus_1CPT,
    inits = inits,
    control = control
  )

}
\author{
Wenping Wang, Mason McComb and Matt Fidler
}
