% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{extractVars}
\alias{extractVars}
\title{Extract all the relevant variables from a set of bootstrapped models}
\usage{
extractVars(fitlist, id = "objf")
}
\arguments{
\item{fitlist}{a list of lists containing information on the multiple bootstrapped models; similar to the output of modelsBootstrap() function}

\item{id}{a character representing the variable of interest: OBJF, AIC, omega, parFixedDf, method, message, warnings}
}
\value{
returns a vector or list across of the variable of interest from all the fits/bootstrapped models
}
\description{
Extract all the relevant variables from a set of bootstrapped models
}
\examples{
extractVars(fitlist, 1) # returns a vector of OBJF values
extractVars(fitlist, 4)  # returns a list of dataframes containing parFixedDf values

}
\author{
Vipul Mann, Matthew Fidler
}
