% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{findLhs}
\alias{findLhs}
\title{Find the assignments in R expression}
\usage{
findLhs(x)
}
\arguments{
\item{x}{R expression}
}
\value{
list of assigned parameters
}
\description{
Find the assignments in R expression
}
\author{
Hadley Wickham and Matthew L. Fidler
}
\keyword{internal}
