% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R, R/nlme_fit.R
\name{focei.eta}
\alias{focei.eta}
\alias{focei.eta.nlmixr_nlme}
\title{Get the FOCEi eta specification for model.}
\usage{
focei.eta(object, ...)

\method{focei.eta}{nlmixr_nlme}(object, ...)
}
\arguments{
\item{object}{Fit object}

\item{...}{Other parameters}
}
\value{
A list for the OMGA list in FOCEi
}
\description{
Get the FOCEi eta specification for model.
}
\author{
Matthew L. Fidler
}
