% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{focei.fit}
\alias{focei.fit}
\title{FOCEI Fit for nlmixr}
\usage{
focei.fit(data, inits, PKpars, diag.xform = c("sqrt", "log", "identity"),
  optim = c("bobyqa", "lbfgsb3", "nlminb", "newuoa", "nlminb", "bobyqa",
  "uobyqa", "praxis", "mma", "slsqp", "lbfgs-nlopt", "tnewton_precond_restart",
  "tnewton_precond", "tnewton", "var2", "var1", "cobyla-nlopt", "bobyqa-nlopt",
  "newuoa-nlopt", "nelder-mead-nlopt", "sbplx"), model = NULL, pred = NULL,
  err = NULL, lower = -Inf, upper = Inf, control = list(),
  calculate.vars = c("pred", "ipred", "ires", "res", "iwres", "wres",
  "cwres"))
}
\arguments{
\item{data}{Data to fit}

\item{inits}{Initilization list}

\item{PKpars}{Pk Parameters}
}
\description{
FOCEI Fit for nlmixr
}
\author{
Matthew L. Fidler
}
