% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R, R/nlme_fit.R
\name{focei.theta}
\alias{focei.theta}
\alias{focei.theta.nlmixrNlme}
\title{Get the FOCEi theta specification for the model}
\usage{
focei.theta(object, uif, ...)

\method{focei.theta}{nlmixrNlme}(object, uif, ...)
}
\arguments{
\item{object}{Fit object}

\item{uif}{User interface function or object}

\item{...}{Other parameters}
}
\value{
Parameter estimates for Theta
}
\description{
Get the FOCEi theta specification for the model
}
