% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foceiFit.R
\name{foceiFit}
\alias{foceiFit}
\title{FOCEi fit}
\usage{
foceiFit(data, inits, PKpars, model = NULL, pred = NULL, err = NULL,
  lower = -Inf, upper = Inf, fixed = NULL, skipCov = NULL,
  control = foceiControl(), thetaNames = NULL, etaNames = NULL,
  etaMat = NULL, ...)
}
\arguments{
\item{data}{Data to fit}

\item{inits}{Initialization list}

\item{PKpars}{Pk Parameters}

\item{model}{The RxODE model to use}

\item{pred}{The Prediction function}

\item{err}{The Error function}

\item{lower}{Lower bounds}

\item{upper}{Upper Bounds}

\item{fixed}{Boolean vector indicating what parameters should be
fixed.}

\item{skipCov}{Boolean vector indicating what parameters should be
fixed when calculating covariances}

\item{control}{FOCEi options Control list}

\item{thetaNames}{Names of the thetas to be used in the final
object}

\item{etaNames}{Eta names to be used in the final object}

\item{etaMat}{Eta matrix for initial estimates or final estimates
of the ETAs.}

\item{...}{Ignored parameters}
}
\value{
A focei fit object

FOCEi fit object
}
\description{
FOCEi fit
}
\examples{


mypar2 <- function ()
{
    ka <- exp(THETA[1] + ETA[1])
    cl <- exp(THETA[2] + ETA[2])
    v  <- exp(THETA[3] + ETA[3])
}

mod <- RxODE({
    d/dt(depot) <- -ka * depot
    d/dt(center) <- ka * depot - cl / v * center
    cp <- center / v
})

pred <- function() cp

err <- function(){
    err <- add(0.1)
}

inits <- list(THTA=c(0.5, -3.2, -1),
              OMGA=list(ETA[1] ~ 1, ETA[2] ~ 2, ETA[3] ~ 1));

fit <- foceiFit(theo_sd, inits, mypar2,mod,pred,err)

}
\author{
Matthew L. Fidler and Wenping Wang

Matthew L. Fidler & Wenping Wang
}
