% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{foceiFitCpp_}
\alias{foceiFitCpp_}
\title{Fit/Evaulate FOCEi}
\usage{
foceiFitCpp_(e)
}
\arguments{
\item{e}{Enviornment}
}
\value{
A focei fit object
}
\description{
This shouldn't be called directly.
}
\keyword{internal}
