% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovar.R
\name{forwardSearch}
\alias{forwardSearch}
\title{Forward covariate search}
\usage{
forwardSearch(covInfo, fit, pVal = 0.05, outputDir, restart = FALSE)
}
\arguments{
\item{covInfo}{a list containing information about each variable-covariate pair}

\item{fit}{an nlmixr 'fit' object}

\item{pVal}{p-value that should be used for selecting covariates in the forward search}

\item{outputDir}{the name of the output directory that stores the covariate search result}

\item{restart}{a boolean that controls if the search should be restarted; default is FALSE}
}
\value{
returns the updated 'fit' object at the end of the forward search and a table of information for all the covaraites tested
}
\description{
Forward covariate search
}
\author{
Vipul Mann, Matthew Fidler
}
