% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.quad.R
\name{gauss.quad}
\alias{gauss.quad}
\title{Sets nodes and weights of Gauss-Hermite quadrature}
\usage{
gauss.quad(n)
}
\arguments{
\item{n}{number of nodes}
}
\value{
a list of nodes and weights of Gauss-Hermite quadrature
}
\description{
Sets nodes and weights of Gauss-Hermite quadrature
}
\examples{
gauss.quad(5)
}
