% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{getBootstrapSummary}
\alias{getBootstrapSummary}
\title{Summarize the bootstrapped fits/models}
\usage{
getBootstrapSummary(fitList, ci = 0.95)
}
\arguments{
\item{fitList}{a list of lists containing information on the multiple bootstrapped models; similar to the output of modelsBootstrap() function}
}
\value{
returns aggregated quantities (mean, median, standard deviation, and variance) as a list for all the quantities
}
\description{
Summarize the bootstrapped fits/models
}
\examples{
getBootstrapSummary(fitlist)

}
\author{
Vipul Mann, Matthew Fidler
}
