% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{getFitMethod}
\alias{getFitMethod}
\title{Get the nlmixr method used for fitting the model}
\usage{
getFitMethod(fit)
}
\arguments{
\item{fit}{the nlmixr fit object}
}
\value{
returns a string representing the method used by nlmixr for fitting the given model
}
\description{
Get the nlmixr method used for fitting the model
}
\examples{
getFitMethod(fit)

}
\author{
Vipul Mann, Matthew Fidler
}
