% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnlmm.R
\name{getOMEGA}
\alias{getOMEGA}
\title{calculate variance-covariance matrix of random effects}
\usage{
getOMEGA(fit)
}
\arguments{
\item{fit}{a gnlmm fit object}
}
\value{
variance-covariance matrix of random effects
}
\description{
calculate variance-covariance matrix of random effects after a gnlmm() fit
}
\examples{
\dontrun{
getOMEGA(fit)
}
}

