% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ini}
\alias{ini}
\title{nlmixr ini block handling}
\usage{
ini(ini, ...)
}
\arguments{
\item{ini}{Ini block or nlmixr model object}

\item{...}{Other arguments parsed by nlmixr}
}
\value{
bounds expression or parsed ui object
}
\description{
nlmixr ini block handling
}
\author{
Matthew L. Fidler
}
