% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovar.R
\name{initializeCovars}
\alias{initializeCovars}
\title{Initializing covariates before estimation}
\usage{
initializeCovars(
  fitobject,
  fstring,
  covNames,
  initialEst,
  initialEstLB,
  initialEstUB
)
}
\arguments{
\item{fitobject}{an nlmixr 'fit' object}

\item{fstring}{a string giving the entire expression for the model function string}

\item{covNames}{a list of covariate names (parameters) that need to be estimates}

\item{initialEst}{the initial estimate for the covariate parameters to be estimated; default is 0}

\item{initialEstLB}{a lower bound for the covariate parameters to be estimated; default is -Inf}

\item{initialEstUB}{an upper bound for the covariate parameters to be estimated; default is Inf}
}
\value{
updated model object with the modified initial values
}
\description{
Initializing covariates before estimation
}
\author{
Vipul Mann, Matthew Fidler
}
