% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instant.stan.extension.R
\name{instant.stan.extension}
\alias{instant.stan.extension}
\title{instant.stan.extension.}
\usage{
instant.stan.extension(ode_str = NULL, covar = NULL)
}
\arguments{
\item{ode_str}{ODE equations in a string}

\item{covar}{a character vector of covariates}
}
\description{
instant.stan.extension
}

