% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgaussian.R
\docType{data}
\name{invgaussian}
\alias{invgaussian}
\title{Inverse Guassian absorption model}
\format{A data frame with 32 rows and 6 columns

\describe{
  \item{time}{Time of observation}
  \item{cp}{Concentration}
}}
\source{
D'Argenio DZ, Schumitzky A, and Wang X (2009). "ADAPT 5 User's Guide: Pharmacokinetic/Pharmacodynamic Systems Analysis Software".
}
\usage{
invgaussian
}
\description{
Inverse Guassian absorption model
}
\keyword{datasets}

