% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acop7-data.R
\docType{data}
\name{iv.1cmt}
\alias{iv.1cmt}
\title{IV 1 compartment model}
\format{An object of class \code{data.frame} with 8873 rows and 15 columns.}
\source{
AcOP7 nlmixr poster
}
\usage{
iv.1cmt
}
\description{
This is a simulated dataset from Schoemaker2016 for comparing
nlmixr to NONMEM.
}
\details{
\describe{
  \item{ID}{Siumlated Subject ID}
  \item{Time}{Simulated Time}
  \item{DV}{Simulated Objective Function}
  \item{LNDV}{Log(DV)}
  \item{MDV}{Missing DV parameter}
  \item{EVID}{Event ID, transformed for nlmixr instead of NONMEM, this includes the infusion rate in the EVID.}
  \item{DOSE}{Dose}
  \item{CMT}{Compartment}
}
}
\keyword{datasets}
