% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem_fit.R
\name{lincmt}
\alias{lincmt}
\title{Parameters for a linear compartment model for SAEM}
\usage{
lincmt(ncmt, oral = T, tlag = F, infusion = F, parameterization = 1)
}
\arguments{
\item{ncmt}{number of compartments}

\item{oral}{logical, whether oral absorption is true}

\item{tlag}{logical, whether lag time is present}

\item{infusion}{logical, whether infusion is true}

\item{parameterization}{type of parameterization, 1=clearance/volume, 2=micro-constants}
}
\value{
parameters for a linear compartment model
}
\description{
Parameters for a linear compartment model using closed-form solutions and the SAEM algorithm.
}
\author{
Wenping Wang
}
