% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{logLik.focei.fit}
\alias{logLik.focei.fit}
\title{Extract log likelihood for focei fit}
\usage{
\method{logLik}{focei.fit}(object, ...)
}
\arguments{
\item{fit}{focei fit object}
}
\value{
logLik object
}
\description{
Extract log likelihood for focei fit
}
\author{
Matthew L. Fidler
}
