% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovar.R
\name{makeHockeyStick}
\alias{makeHockeyStick}
\title{Creating Hockey-stick covariates}
\usage{
makeHockeyStick(data, covariate, varName)
}
\arguments{
\item{data}{a dataframe containing the dataset that needs to be used}

\item{covariate}{the covariate that needs to be converted to hockey-stick; must be present in the data}

\item{varName}{the variable name to which the given covariate is to be added}
}
\value{
a list of updated data with covariates added, an expression that needs to be added to the model expression, the list of covariate names, and the column names corresponding to the hockey-stick covariates
}
\description{
Creating Hockey-stick covariates
}
\author{
Vipul Mann, Matthew Fidler
}
