% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite.R
\docType{data}
\name{metabolite}
\alias{metabolite}
\title{Parent/Metabolite dataset}
\format{
A data frame with 32 rows and 6 columns

\describe{
  \item{time}{Time of observation}
  \item{y1}{Parent Concentration}
  \item{y2}{Metabolite Concentration}
}
}
\source{
D'Argenio DZ, Schumitzky A, and Wang X (2009). "ADAPT 5 User's Guide: Pharmacokinetic/Pharmacodynamic Systems Analysis Software".
}
\usage{
metabolite
}
\description{
Parent/Metabolite dataset
}
\keyword{datasets}
