% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{model}
\alias{model}
\title{nlmixr model block}
\usage{
model(model, ..., .lines = NULL)
}
\arguments{
\item{model}{Model specification}

\item{...}{Other arguments to model object parsed by nlmixr}

\item{.lines}{This is an interal argument when code{model} is
being called recursively and should not be used.}
}
\value{
Parsed UI object
}
\description{
nlmixr model block
}
\author{
Matthew L. Fidler
}
