% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{modelBootstrap}
\alias{modelBootstrap}
\title{Fitting multiple bootstrapped models}
\usage{
modelBootstrap(
  fit,
  numModels = 3,
  nSampIndiv,
  uidCol,
  pvalues = NULL,
  resume = FALSE
)
}
\arguments{
\item{fit}{the nlmixr fit object}

\item{numModels}{an integer giving the number of bootstrapped models to be fit; default value is 3}

\item{nSampIndiv}{an integer specifying the number of samples in each bootstrapped sample; default is the number of unique subjects in the original dataset}

\item{uidCol}{a string representing the unique ID of each subject in the data; default values is 'ID'}

\item{pvalues}{a vector of pvalues indicating the probability of each subject to get selected; default value is NULL}

\item{resume}{a boolean that indicates if a previous session has to be resumed; default value is TRUE}
}
\value{
a list of lists containing the different attributed of the fit object for each of the bootstrapped models
}
\description{
Fitting multiple bootstrapped models
}
\examples{
modelBootstrap(fit)
modelBootstrap(fit, 5)
modelBootstrap(fit, 5, 20)

}
\author{
Vipul Mann, Matthew Fidler
}
