% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n1qn1}
\alias{n1qn1}
\title{n1qn1 optimization}
\usage{
n1qn1(call_eval, call_grad, vars, environment = parent.frame(1), ...,
  epsilon = .Machine$double.eps, max_iterations = 100, nsim = 100,
  imp = 0, invisible = NULL, zm = NULL, restart = FALSE,
  assign = FALSE, print.functions = FALSE)
}
\arguments{
\item{call_eval}{Objective function}

\item{call_grad}{Gradient Function}

\item{vars}{Initial starting point for line search}

\item{environment}{Environment where call_eval/call_grad are
evaluated.}

\item{...}{Ignored additional parameters.}

\item{epsilon}{Precision of estimate}

\item{max_iterations}{Number of iterations}

\item{nsim}{Number of function evaluations}

\item{imp}{Verbosity of messages.}

\item{invisible}{boolean to control if the output of the minimizer
is suppressed.}

\item{zm}{Prior Hessian (in compressed format)}

\item{restart}{Is this an estimation restart?}

\item{assign}{Assign hessian to c.hess in environment environment?
(Default FALSE)}

\item{print.functions}{Boolean to control if the function value
and parameter estimates are echoed every time a function is
called.}
}
\value{
The return value is a list with the following elements:
    \itemize{ \item \code{value} The value at the minimized
    function \item \code{par} The parameter value that minimized
    the function.  \item \code{H} The estimated Hessian at the
    final parameter estimate \item \code{c.hess} Compressed
    Hessian for saving curvature.  }
}
\description{
This is an R port of the n1qn1 optimization procedure in scilab.
}
\author{
C. Lemarechal, Stephen L. Campbell, Jean-Philippe
    Chancelier, Ramine Nikoukhah, Wenping Wang & Matthew L. Fidler
}

