% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\docType{package}
\name{nlmixr-package}
\alias{nlmixr}
\alias{nlmixr-package}
\title{nlmixr fits population PK and PKPD non-linear mixed effects models.}
\description{
nlmixr is an R package for fitting population pharmacokinetic (PK)
and pharmacokinetic-pharmacodynamic (PKPD) models.
importFrom(Rcpp, evalCpp)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nlmixrdevelopment/nlmixr}
}

}
\author{
\strong{Maintainer}: Matthew Fidler \email{matthew.fidler@gmail.com}

Authors:
\itemize{
  \item Yuan Xiong \email{yuan.xiong@gmail.com}
  \item Rik Schoemaker \email{rik.schoemaker@occams.eu}
  \item Justin Wilkins \email{justin.wilkins@occams.eu}
  \item Wenping Wang \email{wwang8198@gmail.com}
}

}
