% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nlmixr.fit}
\alias{nlmixr.fit}
\title{Fit a nlmixr model}
\usage{
\method{nlmixr}{fit}(uif, data, est = "nlme", ...)
}
\arguments{
\item{uif}{Parsed nlmixr model (by \code{nlmixr(mod.fn)}).}

\item{data}{Dataset to estimate}

\item{est}{Estimation method}

\item{...}{Parameters passed to estimation method.}
}
\value{
nlmixr fit object
}
\description{
Fit a nlmixr model
}
\author{
Matthew L. Fidler
}
