% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nlmixr}
\alias{nlmixr}
\alias{nlmixr.function}
\alias{nlmixr.nlmixrUI}
\alias{nlmixr.nlmixr.ui.nlme}
\alias{nlmixr.nlmixr.ui.focei.fit}
\title{nlmixr fits population PK and PKPD non-linear mixed effects models.}
\usage{
nlmixr(object, data, est = "nlme", ...)

\method{nlmixr}{function}(object, data, est = "nlme", ...)

\method{nlmixr}{nlmixrUI}(object, data, est = "nlme", ...)

\method{nlmixr}{nlmixr.ui.nlme}(object, data, est = "nlme", ...)

\method{nlmixr}{nlmixr.ui.focei.fit}(object, data, est = "nlme", ...)
}
\arguments{
\item{object}{Fitted object or}

\item{data}{Data for fit.}

\item{est}{Estimation routine.}

\item{...}{Other parameters}
}
\value{
Either a nlmixr model or a nlmixr fit object
}
\description{
nlmixr is an R package for fitting population pharmacokinetic (PK)
and pharmacokinetic-pharmacodynamic (PKPD) models.
}
\details{
The nlmixr generalized function allows common access to the nlmixr
estimation routines.
}
\author{
Matthew L. Fidler
}
