% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{nlmixrBoundsParser}
\alias{nlmixrBoundsParser}
\alias{nlmixrBoundsParser.(}
\title{Functions to assist with setting initial conditions and boundaries}
\usage{
nlmixrBoundsParser(x)

\method{nlmixrBoundsParser}{`(`}(x)
}
\arguments{
\item{x}{the object to attempt extraction from}
}
\value{
A list with how the object will be used
}
\description{
These functions are not intended to be called by a user.  They are intended
to be internal to nlmixr
}
\section{Methods (by class)}{
\itemize{
\item \code{(}: For function bodies and similar.
}}

\seealso{
Other nlmixrBounds: 
\code{\link{nlmixrBounds}()}
}
\concept{nlmixrBounds}
