% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{nlmixrBoundsParserTheta}
\alias{nlmixrBoundsParserTheta}
\title{Forms of fixed that are allowed are:}
\usage{
nlmixrBoundsParserTheta(x, currentData)
}
\description{
* \code{fixed(1)}
* \code{fixed(1, 2)}
* \code{fixed(1, 2, 3)}
* \code{c(1, fixed)}
* \code{c(1, 2, fixed)}
* \code{c(1, 2, 3, fixed)}
* \code{c(1, fixed(2))}
* \code{c(1, fixed(2), 3)}
}
\details{
Where 'fixed' can be 'FIX', 'FIXED', 'fix', or 'fixed'.
}
