% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\docType{data}
\name{nlmixrBoundsTemplate}
\alias{nlmixrBoundsTemplate}
\title{This is defines the columns and classes for the nlmixrBounds data.frame.
(Done here to make testing easier and ensure consistency across functions.)}
\format{
An object of class \code{data.frame} with 1 rows and 11 columns.
}
\usage{
nlmixrBoundsTemplate
}
\description{
This is defines the columns and classes for the nlmixrBounds data.frame.
(Done here to make testing easier and ensure consistency across functions.)
}
\keyword{datasets}
