% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlmixrDynmodelConvert}
\alias{nlmixrDynmodelConvert}
\title{Converting nlmixr objects to dynmodel objects}
\usage{
nlmixrDynmodelConvert(.nmf)
}
\arguments{
\item{.nmf}{nlmixr object}
}
\value{
list containing inputs for the dynmodel()
\itemize{
\item \code{$fixPars} - fixed parameters defined as \code{fixed()} in the nlmixr object
\item \code{$sigma} - error model parameters
\item \code{$inits} - initial estimates for parameters in the model
\item \code{$lower} - lower boundaries for estimated parameters
\item \code{$upper} - upper boundaries for estimated parameters
\item \code{$system} - RxODE object that defines the structural model
\item \code{$model} - error model
}
}
\description{
Convert nlmixr Objects to dynmodel objects for use in fitting non-population dynamic models
}
\author{
Mason McComb and Matt Fidler
}
