% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nlmixrEst}
\alias{nlmixrEst}
\alias{nlmixrEst.saem}
\alias{nlmixrEst.nlme}
\alias{nlmixrEst.nlme.mu}
\alias{nlmixrEst.nlme.mu.cov}
\alias{nlmixrEst.nlme.free}
\alias{nlmixrEst.posthoc}
\alias{nlmixrEst.focei}
\alias{nlmixrEst.foce}
\alias{nlmixrEst.fo}
\alias{nlmixrEst.foi}
\alias{nlmixrEst.dynmodel}
\alias{nlmixrEst.nlmixrEst}
\title{Generic for nlmixr estimation methods}
\usage{
nlmixrEst(env, ...)

\method{nlmixrEst}{saem}(env, ...)

\method{nlmixrEst}{nlme}(env, ...)

\method{nlmixrEst}{nlme.mu}(env, ...)

\method{nlmixrEst}{nlme.mu.cov}(env, ...)

\method{nlmixrEst}{nlme.free}(env, ...)

\method{nlmixrEst}{posthoc}(env, ...)

\method{nlmixrEst}{focei}(env, ...)

\method{nlmixrEst}{foce}(env, ...)

\method{nlmixrEst}{fo}(env, ...)

\method{nlmixrEst}{foi}(env, ...)

\method{nlmixrEst}{posthoc}(env, ...)

\method{nlmixrEst}{dynmodel}(env, ...)

\method{nlmixrEst}{nlmixrEst}(env, ...)
}
\arguments{
\item{env}{Environment for nlmixr estimation routines}

\item{...}{Extra arguments sent to estimation routine}
}
\value{
nlmixr estimation object
}
\description{
Generic for nlmixr estimation methods
}
\details{
This is a S3 generic that allows others to use the nlmixr
  environment to do their own estimation routines
}
\author{
Matthew Fidler
}
