% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nlmixrLogo}
\alias{nlmixrLogo}
\title{Messages the nlmixr logo...}
\usage{
nlmixrLogo(str = "", version = sessionInfo()$otherPkgs$nlmixr$Version)
}
\arguments{
\item{str}{String to print}

\item{version}{Version information (by default use package version)}
}
\description{
Messages the nlmixr logo...
}
\author{
Matthew L. Fidler
}
