% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foceiFit.R
\name{nlmixrPosthoc}
\alias{nlmixrPosthoc}
\title{Get a posthoc estimate of x}
\usage{
nlmixrPosthoc(x, ...)
}
\arguments{
\item{x}{nlmixr object}

\item{...}{other arguments}
}
\value{
nlmixr fit object with possibly a new set of estimates
}
\description{
Get a posthoc estimate of x
}
\keyword{internal}
