% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{nlmixrPred}
\alias{nlmixrPred}
\alias{predict.nlmixrFitData}
\title{Predict a nlmixr solved system}
\usage{
nlmixrPred(object, ..., ipred = FALSE)

\method{predict}{nlmixrFitData}(object, ...)
}
\arguments{
\item{ipred}{Flag to calculate individual predictions. When
\code{ipred} is \code{TRUE}, calculate individual predictions.
When \code{ipred} is \code{FALSE}, set calculate typical population predations.
When \code{ipred} is \code{NA}, calculate both individual and
population predictions.}
}
\description{
Predict a nlmixr solved system
}
