% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlmixrPrint}
\alias{nlmixrPrint}
\title{Print x using the message facility}
\usage{
nlmixrPrint(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{Other things output}
}
\description{
This allows the suppressMessages to work on print functions.  This
captures the output function sends it through the message routine.
}
\details{
catpureOutput was used since it is much faster than the internal
capture.output see https://www.r-bloggers.com/performance-captureoutput-is-much-faster-than-capture-output/
}
\author{
Matthew L. Fidler
}
\keyword{internal}
