% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlmixrTest}
\alias{nlmixrTest}
\title{nlmixTest function for testing}
\usage{
nlmixrTest(expr, silent = .isTestthat(), test = "cran")
}
\arguments{
\item{expr}{Expression for testing}

\item{silent}{Boolean for testing}

\item{test}{this represents the test group of the test}
}
\description{
nlmixTest function for testing
}
\author{
Matthew Fidler
}
