% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.dynmodelfun}
\alias{nlmixrUI.dynmodelfun}
\title{Return dynmodel variable translation function}
\usage{
nlmixrUI.dynmodelfun(object)
}
\arguments{
\item{object}{nlmixr ui object}
}
\value{
nlmixr dynmodel translation
}
\description{
Return dynmodel variable translation function
}
\author{
Matthew Fidler
}
