% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.eta.names}
\alias{nlmixrUI.eta.names}
\title{Get ETA names}
\usage{
nlmixrUI.eta.names(obj)
}
\arguments{
\item{obj}{UI object}
}
\value{
ETA names
}
\description{
Get ETA names
}
\author{
Matthew L. Fidler
}
