% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.focei.fixed}
\alias{nlmixrUI.focei.fixed}
\title{Get parameters that are fixed}
\usage{
nlmixrUI.focei.fixed(obj)
}
\arguments{
\item{obj}{UI object}
}
\value{
logical vector of fixed THETA parameters
}
\description{
Get parameters that are fixed
}
\author{
Matthew L. Fidler
}
