% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.focei.inits}
\alias{nlmixrUI.focei.inits}
\title{Get the FOCEi initilizations}
\usage{
nlmixrUI.focei.inits(obj)
}
\arguments{
\item{obj}{UI object}
}
\value{
list with FOCEi style initilizations
}
\description{
Get the FOCEi initilizations
}
\author{
Matthew L. Fidler
}

