% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.nlme.specs}
\alias{nlmixrUI.nlme.specs}
\title{Create the nlme specs list for nlmixr nlme solving}
\usage{
nlmixrUI.nlme.specs(object, mu.ref = FALSE)
}
\arguments{
\item{object}{UI object}

\item{mu.ref}{Use mu-referencing for nlme.}
}
\value{
specs list for nlme
}
\description{
Create the nlme specs list for nlmixr nlme solving
}
\author{
Matthew L. Fidler
}
