% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI}
\alias{nlmixrUI}
\title{Prepares the UI function and returns a list.}
\usage{
nlmixrUI(fun)
}
\arguments{
\item{fun}{UI function}
}
\value{
nlmixr UI function
}
\description{
Prepares the UI function and returns a list.
}
\author{
Matthew L. Fidler
}
