% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.rxode.pred}
\alias{nlmixrUI.rxode.pred}
\title{Return RxODE model with predictions appended}
\usage{
nlmixrUI.rxode.pred(object)
}
\arguments{
\item{object}{UI object}
}
\value{
String or NULL if RxODE is not specified by UI.
}
\description{
Return RxODE model with predictions appended
}
\author{
Matthew L. Fidler
}
