% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.saem.cres}
\alias{nlmixrUI.saem.cres}
\title{Get initial estimate for bres SAEM.}
\usage{
nlmixrUI.saem.cres(obj)
}
\arguments{
\item{obj}{UI model}
}
\value{
SAEM model$ares spec
}
\description{
Get initial estimate for bres SAEM.
}
\author{
Matthew L. Fidler
}
