% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.saem.distribution}
\alias{nlmixrUI.saem.distribution}
\title{Get SAEM distribution}
\usage{
nlmixrUI.saem.distribution(obj)
}
\arguments{
\item{obj}{UI object}
}
\value{
Character of distribution
}
\description{
Get SAEM distribution
}
\author{
Matthew L. Fidler
}
