% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.saem.init.omega}
\alias{nlmixrUI.saem.init.omega}
\title{SAEM's init$omega}
\usage{
nlmixrUI.saem.init.omega(obj)
}
\arguments{
\item{obj}{nlmixr UI object}
}
\value{
Return initial matrix
}
\description{
SAEM's init$omega
}
\author{
Matthew L. Fidler
}
