% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.saem.init}
\alias{nlmixrUI.saem.init}
\title{Get saem initilization list}
\usage{
nlmixrUI.saem.init(obj)
}
\arguments{
\item{obj}{nlmixr UI object}
}
\value{
Return SAEM inits list.
}
\description{
Get saem initilization list
}
\author{
Matthew L. Fidler
}

