% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.saem.model}
\alias{nlmixrUI.saem.model}
\title{Generate SAEM model list}
\usage{
nlmixrUI.saem.model(obj)
}
\arguments{
\item{obj}{nlmixr UI object}
}
\value{
SAEM model list
}
\description{
Generate SAEM model list
}
\author{
Matthew L. Fidler
}
