% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.saem.res.mod}
\alias{nlmixrUI.saem.res.mod}
\title{Get the SAEM model$res.mod code}
\usage{
nlmixrUI.saem.res.mod(obj)
}
\arguments{
\item{obj}{UI model}
}
\value{
SAEM model$res.mod spec
}
\description{
Get the SAEM model$res.mod code
}
\author{
Matthew L. Fidler
}
