% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.saem.rx1}
\alias{nlmixrUI.saem.rx1}
\title{Return RxODE model with predictions appended}
\usage{
nlmixrUI.saem.rx1(object)
}
\arguments{
\item{object}{UI object}
}
\value{
Combined focei model text for RxODE
}
\description{
Return RxODE model with predictions appended
}
\author{
Matthew L. Fidler
}
