% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.saem.theta.name}
\alias{nlmixrUI.saem.theta.name}
\title{Get THETA names for nlmixr's SAEM}
\usage{
nlmixrUI.saem.theta.name(uif)
}
\arguments{
\item{uif}{nlmixr UI object}
}
\value{
SAEM theta names
}
\description{
Get THETA names for nlmixr's SAEM
}
\author{
Matthew L. Fidler
}

