% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUILinCmt}
\alias{nlmixrUILinCmt}
\title{Build linear solved information based on defined parameters.}
\usage{
nlmixrUILinCmt(lhs)
}
\arguments{
\item{lhs}{List of defined parameters}
}
\value{
List containing the translation parmateres (extra lines)
    to give what nlmixr expects, and model properties
    (parameterization, ncmt, oral, and tlag).  The infusion needs
    to be guessed from the data and is not included in this function's output
}
\description{
Build linear solved information based on defined parameters.
}
\author{
Matthew L. Fidler
}

