% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{nlmixrValidate}
\alias{nlmixrValidate}
\alias{nmTest}
\title{Validate nlmixr}
\usage{
nlmixrValidate(type = NULL, check = FALSE)

nmTest(type = NULL, check = FALSE)
}
\arguments{
\item{type}{of test to be run}

\item{check}{Use devtools::check to run checks}
}
\description{
This allows easy vaildation/qualification of nlmixr by running the
testing suite on your system.
}
\author{
Matthew L. Fidler
}
