% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nlmixrVersion}
\alias{nlmixrVersion}
\title{Display nlmixr's version}
\usage{
nlmixrVersion()
}
\value{
Nothing, called for its side effects
}
\description{
Display nlmixr's version
}
\author{
Matthew L. Fidler
}
