% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConvert.R
\name{nmDataConvert}
\alias{nmDataConvert}
\title{Convert NONMEM/Monolix data to RxODE/nlmixr style data}
\usage{
nmDataConvert(nonmem.data)
}
\arguments{
\item{nonmem.data}{NONMEM dataset}
}
\value{
RxODE compatible data.
}
\description{
Convert NONMEM/Monolix data to RxODE/nlmixr style data
}
\author{
Yuan Xiong and Matthew L. Fidler
}

