% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable.R
\name{nmLst}
\alias{nmLst}
\title{Create a large output based on a nlmixr fit}
\usage{
nmLst(x, lst = NULL)
}
\arguments{
\item{x}{nlmixr fit object}

\item{lst}{Listing file.  If not specified, it is determined by
the day and the model/R-object name.  If it is specified as
\code{NULL} the listing output is displayed on the screen.}
}
\value{
invisibly returns fit
}
\description{
Create a large output based on a nlmixr fit
}
\author{
Matthew Fidler
}
