% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nmsimplex}
\alias{nmsimplex}
\title{Nelder-Mead simplex search}
\usage{
nmsimplex(start, fr, rho = NULL, control = list())
}
\arguments{
\item{start}{initials}

\item{fr}{objective function}

\item{rho}{evaluation environment}

\item{control}{additional optimization options}
}
\value{
a list of ...
}
\description{
Nelder-Mead simplex search
}
