% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem_fit.R
\name{nmxInclude}
\alias{nmxInclude}
\title{Get a list of directories for inclusion}
\usage{
nmxInclude(pkg = "nlmixr")
}
\arguments{
\item{pkg}{a string or list of string for packages to be included.}
}
\value{
An inclusion string for Makevars
}
\description{
Get a list of directories for inclusion
}
\author{
Matthew L. Fidler
}
\keyword{internal}

