% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{nobs.focei.fit}
\alias{nobs.focei.fit}
\title{Extract the number of observations for the focei fit}
\usage{
\method{nobs}{focei.fit}(object, ...)
}
\arguments{
\item{object}{focei fit object}

\item{...}{other parameters (ignored)}
}
\value{
Number of observations
}
\description{
Extract the number of observations for the focei fit
}
\author{
Matthew L. Fidler
}

