% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{par.hist}
\alias{par.hist}
\alias{par.hist.nlmixr.ui.saem}
\alias{par.hist.nlmixr.ui.focei.fit}
\title{Parameter history extraction}
\usage{
par.hist(x, stacked = FALSE, ...)

\method{par.hist}{nlmixr.ui.saem}(x, stacked = FALSE, ...)

\method{par.hist}{nlmixr.ui.focei.fit}(x, stacked = FALSE, ...)
}
\arguments{
\item{x}{object to extract parameter history from}

\item{stacked}{Should the data frame be stacked (default FALSE)}

\item{...}{Other parameters (currently ignored)}
}
\value{
A data frame with the parameter history.  NULL if no
    information is available
}
\description{
Parameter history extraction
}
\author{
Matthew L. Fidler & Wenping Wang
}
