% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynmodel.R
\name{plot.dyn.mcmc}
\alias{plot.dyn.mcmc}
\title{Plot of a non-population dynamic model fit using mcmc}
\usage{
\method{plot}{dyn.mcmc}(x, ...)
}
\arguments{
\item{x}{a dynmodel fit object}

\item{...}{additional arguments}
}
\value{
nothing, called to produce goodness of fits
}
\description{
Plot of a non-population dynamic model fit using mcmc
}
