% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R, R/utils_cmt_ode12.R
\name{plot.focei.fit}
\alias{plot.focei.fit}
\alias{plot.focei.fit}
\title{Plot a focei.fit plot}
\usage{
\method{plot}{focei.fit}(fit)

\method{plot}{focei.fit}(fit)
}
\arguments{
\item{x}{a focei fit object}

\item{...}{additional arguments}

\item{x}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
Plot some standard goodness of fit plots for the focei fitted object

Print a generalized non-linear mixed effect model fit
}
\author{
Wenping Wang & Matthew Fidler
}
