% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem_fit_aux.R
\name{plot.saemFit}
\alias{plot.saemFit}
\title{Plot an SAEM model fit}
\usage{
\method{plot}{saemFit}(x, ...)
}
\arguments{
\item{x}{a saemFit object}

\item{...}{others}
}
\value{
a list
}
\description{
Plot an SAEM model fit
}
