% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preCondInv}
\alias{preCondInv}
\title{Calculate the inverse preconditioning matrix}
\usage{
preCondInv(Rin)
}
\arguments{
\item{Rin}{The R matrix input}
}
\value{
The inverse preconditioning matrix
}
\description{
Calculate the inverse preconditioning matrix
}
