% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dyn.ID}
\alias{print.dyn.ID}
\title{Print a non-population dynamic model fit object}
\usage{
\method{print}{dyn.ID}(x, ...)
}
\arguments{
\item{x}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
Print a non-population dynamic model fit object
}

