% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R, R/utils_cmt_ode12.R
\name{print.focei.fit}
\alias{print.focei.fit}
\alias{print.focei.fit}
\title{Print a focei fit}
\usage{
\method{print}{focei.fit}(fit)

\method{print}{focei.fit}(fit)
}
\arguments{
\item{x}{a focei model fit}

\item{...}{additional arguments}

\item{x}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
Print a first-order conditional non-linear mixed effect model with
interaction fit

Print a generalized non-linear mixed effect model fit
}
