% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnlmm.R
\name{print.gnlmm.fit}
\alias{print.gnlmm.fit}
\title{Print a gnlmm fit}
\usage{
\method{print}{gnlmm.fit}(x, ...)
}
\arguments{
\item{x}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
Print a generalized non-linear mixed effect model fit
}

