% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{print.nlmixrUI}
\alias{print.nlmixrUI}
\title{Print UI function}
\usage{
\method{print}{nlmixrUI}(x, ...)
}
\arguments{
\item{x}{UI function}

\item{...}{other arguments}
}
\value{
original object (invisibly)
}
\description{
Print UI function
}
\author{
Matthew L. Fidler
}
