% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem_fit.R
\name{print.saemFit}
\alias{print.saemFit}
\title{print a saem fit}
\usage{
\method{print}{saemFit}(x, ...)
}
\arguments{
\item{x}{a saemFit object}

\item{...}{others}
}
\value{
a list
}
\description{
print a saem fit
}

