% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{printBootstrapSummary}
\alias{printBootstrapSummary}
\title{Print a well-formatted summary for the bootstrap models}
\usage{
printBootstrapSummary(x, fitObj)
}
\arguments{
\item{x}{the summary object returned by the getBootstrapSummary() function}
}
\description{
Print a well-formatted summary for the bootstrap models
}
\author{
Vipul Mann, Matthew Fidler
}
