% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovar.R
\name{removeCovMultiple}
\alias{removeCovMultiple}
\title{Removing multiple covariates}
\usage{
removeCovMultiple(covInfo, fitobject)
}
\arguments{
\item{covInfo}{a list containing information about each variable-covariate pair}

\item{fitobject}{an nlmixr 'fit' object}
}
\value{
a list with the updated fit object, the variable-covariate pair string, and the parameter names for the corresponding covaraites removed
}
\description{
Removing multiple covariates
}
\author{
Vipul Mann, Matthew Fidler
}
