% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{residuals.focei.fit}
\alias{residuals.focei.fit}
\title{Extract residuals from the FOCEI fit}
\usage{
\method{residuals}{focei.fit}(object, ..., type = c("ires", "res", "iwres",
  "wres", "cwres", "cpred", "cres"), etas = NULL)
}
\arguments{
\item{object}{focei.fit object}

\item{...}{Additional arguments}

\item{type}{Residuals type fitted.}

\item{etas}{ETAs matrix to use for the calculation.}
}
\value{
residuals
}
\description{
Extract residuals from the FOCEI fit
}
\author{
Matthew L. Fidler
}
