% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foceiFit.R
\name{residuals.nlmixrFitData}
\alias{residuals.nlmixrFitData}
\title{Extract residuals from the FOCEI fit}
\usage{
\method{residuals}{nlmixrFitData}(object, ..., type = c("ires", "res",
  "iwres", "wres", "cwres", "cpred", "cres"))
}
\arguments{
\item{object}{focei.fit object}

\item{...}{Additional arguments}

\item{type}{Residuals type fitted.}
}
\value{
residuals
}
\description{
Extract residuals from the FOCEI fit
}
\author{
Matthew L. Fidler
}
