% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{rxFoceiEta}
\alias{rxFoceiEta}
\title{FOCEI ETA setup,}
\usage{
rxFoceiEta(object, ..., dv, eta, env, nonmem = FALSE)
}
\arguments{
\item{object}{RxODE object}

\item{...}{values sent to rxFoceiEtaSetup}

\item{dv}{dependent variable}

\item{eta}{eta values.}

\item{env}{Environment to use, instead of rxFoceiEtaSetup environment}

\item{nonmem}{Match NONMEMs approximation of the hessian.}
}
\value{
environment of solved information.
}
\description{
This is basically for testing
}
\author{
Matthew L. Fidler
}
\keyword{internal}

