% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{rxFoceiGrad}
\alias{rxFoceiGrad}
\title{Calculate the Gradient for an individual}
\usage{
rxFoceiGrad(object, ret, ..., theta, eta = NULL, dv,
  numDeriv.method = "Richardson")
}
\arguments{
\item{object}{focei object}

\item{ret}{individual return object to add "grad" attribute to...}

\item{...}{values sent to rxFoceiEtaSetup}

\item{eta}{eta values.}

\item{dv}{dependent variable}
}
\description{
This can be based on a combination of numerical differention and
symbolic differentiation or purely numerical differentiation.
}
\keyword{internal}
