% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{rxFoceiLp}
\alias{rxFoceiLp}
\title{Get -Grad for individual}
\usage{
rxFoceiLp(object, ..., dv, eta)
}
\arguments{
\item{object}{RxODE object}

\item{...}{values sent to rxFoceiEtaSetup}

\item{dv}{dependent variable}

\item{eta}{eta values.}
}
\value{
-grad
}
\description{
For testing
}
\author{
Matthew L. Fidler
}
\keyword{internal}

