% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxGrad}
\alias{rxGrad}
\title{Returns the gradient for FOCEi environment}
\usage{
rxGrad(rho)
}
\arguments{
\item{rho}{Environment to calculate the gradient for...}
}
\description{
Returns the gradient for FOCEi environment
}
\keyword{internal}

