% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxHessian}
\alias{rxHessian}
\title{Get the Hessian for the environment}
\usage{
rxHessian(rho)
}
\arguments{
\item{rho}{environment}
}
\description{
Get the Hessian for the environment
}
\keyword{internal}

