% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RxODE_focei_eta}
\alias{RxODE_focei_eta}
\title{Get the Function pointers for the LBJ'S routine}
\usage{
RxODE_focei_eta(fstr)
}
\arguments{
\item{fstr}{a string of the function pointer to return.  "lik" for likelihood and "lp" for likelihood gradient.}
}
\description{
Get the Function pointers for the LBJ'S routine
}
\keyword{internal}

