% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RxODE_focei_eta_lp}
\alias{RxODE_focei_eta_lp}
\title{Get the likelihood slope for ETA}
\usage{
RxODE_focei_eta_lp(sexp_eta, sexp_rho)
}
\arguments{
\item{sexp_eta}{ETA for likelihood}

\item{sexp_rho}{Environment with solving options}
}
\description{
Get the likelihood slope for ETA
}
\keyword{internal}

