% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RxODE_focei_finalize_llik}
\alias{RxODE_focei_finalize_llik}
\title{Finalize likelihood environment}
\usage{
RxODE_focei_finalize_llik(rho)
}
\arguments{
\item{rho}{Environment to finalize.  Returns an individual likelihood.}
}
\description{
Finalize likelihood environment
}
\keyword{internal}

