% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxUpdateEtas}
\alias{rxUpdateEtas}
\title{Update ETAs based on d(eta)/d(theta)}
\usage{
rxUpdateEtas(DnDhS, DhS, initS, acceptNS)
}
\arguments{
\item{DnDhS}{This is the d(eta)/d(theta) list where there is a
d(eta)/d(theta) matrix for each subject}

\item{DhS}{This is the change in theta observed between steps.}

\item{initS}{This is the ETA initial condition matrix}

\item{acceptNS}{Acceptance criteria for the new eta.  |eta| < acceptNS for the new eta
to be accepted.}
}
\description{
This updates the ETA initial estimates based on the knowledge of d(eta)/d(theta)
}
\keyword{internal}
