% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem_fit.R
\name{saem.cleanup}
\alias{saem.cleanup}
\title{Cleanup saem_fit environment by removing dll after the object is no logner used by R.}
\usage{
\method{saem}{cleanup}(env)
}
\arguments{
\item{env}{Environment where cleanup needs to occur.}
}
\description{
Cleanup saem_fit environment by removing dll after the object is no logner used by R.
}
\author{
Matthew L. Fidler
}
