% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{saemControl}
\alias{saemControl}
\title{Control Options for SAEM}
\usage{
saemControl(seed = 99, sa.iter = 200, em.iter = 300, nmc = 3,
  nu = c(2, 2, 2), atol = 1e-08, rtol = 1e-06, stiff = TRUE,
  transit_abs = FALSE, print = 1, ...)
}
\arguments{
\item{seed}{Random Seed for SAEM step.  (Needs to be set for
reproducibility.)  By default this is 99.}

\item{sa.iter}{Number of iterations in the Stochastic
Approximation (SA) Step}

\item{em.iter}{Number of iterations in the
Expectation-Maximization (EM) Step}

\item{nmc}{Number of Markov Chains?}

\item{nu}{Nu! A new parameter....?}

\item{atol}{a numeric absolute tolerance (1e-08 by default);}

\item{rtol}{a numeric relative tolerance (1e-06 by default).}

\item{stiff}{a logical (\code{TRUE} by default) indicating whether
    the ODE system is stiff or not.

    For stiff ODE sytems (\code{stiff = TRUE}), \code{RxODE} uses the
    LSODA (Livermore Solver for Ordinary Differential Equations)
    Fortran package, which implements an automatic method switching
    for stiff and non-stiff problems along the integration
    interval, authored by Hindmarsh and Petzold (2003).

    For non-stiff systems (\code{stiff = FALSE}), \code{RxODE} uses
    DOP853, an explicit Runge-Kutta method of order 8(5, 3) of
    Dormand and Prince as implemented in C by Hairer and Wanner
    (1993).}

\item{transit_abs}{boolean indicating if this is a transit
compartment absorption}

\item{print}{The number it iterations that are completed before
anything is printed to the console.  By default, this is 1.}

\item{...}{Other arguments to control SAEM.}
}
\value{
List of options to be used in \code{\link{nlmixr}} fit for
    SAEM.
}
\description{
Control Options for SAEM
}
\author{
Wenping Wang & Matthew L. Fidler
}
