% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{sampling}
\alias{sampling}
\title{Bootstrap sampling from a given dataframe}
\usage{
sampling(data, nsamp, uid_colname, pvalues = NULL)
}
\arguments{
\item{data}{the original dataframe object to sample from for bootstrapping}

\item{nsamp}{an integer specifying the number of samples in each bootstrapped sample; default is the number of unique subjects in the original dataset}

\item{uid_colname}{a string representing the unique ID of each subject in the data; default values is 'ID'}

\item{pvalues}{a vector of pvalues indicating the probability of each subject to get selected; default value is NULL}
}
\value{
returns a bootstrap sampled dataframe object
}
\description{
Bootstrap sampling from a given dataframe
}
\examples{
sampling(data)
sampling(data, 10)

}
\author{
Vipul Mann, Matthew Fidler
}
