% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{setCov}
\alias{setCov}
\title{Set the covariance type based on prior calculated covariances}
\usage{
setCov(fit, method)
}
\arguments{
\item{fit}{nlmixr fit}

\item{method}{covariance method}
}
\value{
Fit object with covariance updated
}
\description{
Set the covariance type based on prior calculated covariances
}
\author{
Matt Fidler
}
