% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofv.R
\name{setOfv}
\alias{setOfv}
\alias{getOfvType}
\title{Set/get Objective function type for a nlmixr object}
\usage{
setOfv(x, type)

getOfvType(x)
}
\arguments{
\item{x}{nlmixr fit object}

\item{type}{Type of objective function to use for AIC, BIC, and
$objective}
}
\value{
Nothing
}
\description{
Set/get Objective function type for a nlmixr object
}
\author{
Matthew L. Fidler
}
