% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{simulate.focei.fit}
\alias{simulate.focei.fit}
\title{Simulate response based on FOCEi model's datathe fitted object's dataset}
\usage{
\method{simulate}{focei.fit}(object, nsim = 1, seed = NULL, ...)
}
\value{
New dataset based on original dataset
}
\description{
Simulate response based on FOCEi model's datathe fitted object's dataset
}
\author{
Matthew L. Fidler
}
