% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynmodel.R
\name{summary.dyn.ID}
\alias{summary.dyn.ID}
\title{Summary of a non-population dynamic model fit}
\usage{
\method{summary}{dyn.ID}(object, ...)
}
\arguments{
\item{object}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
Summary of a non-population dynamic model fit
}
