% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theo_sd.R
\docType{data}
\name{theo_sd}
\alias{theo_sd}
\title{Multiple dose Theophiline PK data}
\format{A data frame with 144 rows by 7 columns

\describe{
  \item{ID}{Subject ID}
  \item{TIME}{Time (hrs)}
  \item{DV}{Dependant Variable, Theophiline Concentration}
  \item{AMT}{Dose Amount}
  \item{EVID}{RxODE/nlmixr event ID (not NONMEM's)}
  \item{CMT}{Compartment Number}
  \item{WT}{Weight (kg)}
}}
\source{
NONMEM/nlme
}
\usage{
theo_sd
}
\description{
This datastart is  the day 1 concentrations of the
Theophiline data that is included in the nlme/NONMEM.
}
\keyword{datasets}

