% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit.R
\name{VarCorr.nlmixrNlme}
\alias{VarCorr.nlmixrNlme}
\title{Return VarCorr for nlmixr nlme}
\usage{
\method{VarCorr}{nlmixrNlme}(x, sigma = NULL, ...)
}
\arguments{
\item{x}{a fitted model object, usually an object inheriting from
    class \code{"\link[nlme]{lme}"}.}

\item{sigma}{an optional numeric value used as a multiplier for the
    standard deviations.  The default is \code{x$sigma} or \code{1}
    depending on \code{\link{class}(x)}.}

\item{...}{further optional arguments passed to other methods (none
    for the methods documented here).}
}
\description{
This returns a numeric matrix instead of character matrix
}
\author{
Matthew L. Fidler
}
