% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{vpc}
\alias{vpc}
\title{Visual predictive check (VPC)}
\usage{
vpc(fit, nsim = 100, condition = NULL)
}
\arguments{
\item{fit}{nlme fit object}

\item{nsim}{number of simulations}

\item{condition}{conditional variable}
}
\description{
Do visual predictive check (VPC) plots for nlme-base non-linear mixed effect models
}
\examples{
\dontrun{
vpc(fit, nsim = 100, condition = NULL)
}
}
