% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.saemFit.R
\name{vpc_saemFit}
\alias{vpc_saemFit}
\alias{vpc.saemFit}
\title{VPC for nlmixr saemFit objects}
\usage{
vpc_saemFit(fit, dat, nsim = 100, by = NULL, ...)

\method{vpc}{saemFit}(sim, ...)
}
\arguments{
\item{fit}{saemFit object}

\item{dat}{Data to augment the saemFit vpc simulation}

\item{nsim}{Number of simulations for the VPC}

\item{by}{Variables to condition the VPC}

\item{...}{Other arguments sent to \code{\link[vpc:vpc]{vpc_vpc}}}
}
\value{
vpc object from the \code{\link[vpc:vpc]{vpc_vpc}} package
}
\description{
VPC for nlmixr saemFit objects
}
\author{
Wenping Wang
}
