% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.ui.R
\name{vpc_ui}
\alias{vpc_ui}
\alias{vpc.nlmixrFitData}
\alias{vpc.nlmixrVpc}
\alias{vpc.ui}
\title{VPC based on ui model}
\usage{
vpc_ui(
  fit,
  data = NULL,
  n = 100,
  bins = "jenks",
  n_bins = "auto",
  bin_mid = "mean",
  show = NULL,
  stratify = NULL,
  pred_corr = FALSE,
  pred_corr_lower_bnd = 0,
  pi = c(0.05, 0.95),
  ci = c(0.05, 0.95),
  uloq = NULL,
  lloq = NULL,
  log_y = FALSE,
  log_y_min = 0.001,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  smooth = TRUE,
  vpc_theme = NULL,
  facet = "wrap",
  labeller = NULL,
  vpcdb = FALSE,
  verbose = FALSE,
  ...
)

\method{vpc}{nlmixrFitData}(sim, ...)

\method{vpc}{nlmixrVpc}(sim, ...)

\method{vpc}{ui}(sim, ...)
}
\arguments{
\item{fit}{nlmixr fit object}

\item{data}{this is the data to use to augment the VPC fit.  By
default is the fitted data, (can be retrieved by
\code{\link[nlme]{getData}}), but it can be changed by specifying
this argument.}

\item{n}{Number of VPC simulations.  By default 100}

\item{...}{Args sent to \code{\link[RxODE]{rxSolve}}}
}
\value{
Simulated dataset (invisibly)
}
\description{
VPC based on ui model
}
\author{
Matthew L. Fidler
}
