% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Wang2007.R
\docType{data}
\name{Wang2007}
\alias{Wang2007}
\title{Simulated Data Set for comparing objective functions}
\format{
A data frame with 20 rows and 3 columns

\describe{
  \item{ID}{Simulated Subject ID}
  \item{Time}{Simulated Time}
  \item{Y}{Simulated Value}
}
}
\source{
Table 1 from Wang, Y \emph{Derivation of Various NONMEM estimation methods}. J Pharmacokinet Pharmacodyn (2007) 34:575-593.
}
\usage{
Wang2007
}
\description{
This is a simulated dataset from Wang2007 where various NONMEM
estimation methods (Laplace FO, FOCE with and without
interaction) are described.
}
\seealso{
Other nlmixr datasets: 
\code{\link{Bolus_1CPTMM}},
\code{\link{Bolus_1CPT}},
\code{\link{Bolus_2CPTMM}},
\code{\link{Bolus_2CPT}},
\code{\link{Infusion_1CPT}},
\code{\link{Oral_1CPT}},
\code{\link{pheno_sd}},
\code{\link{rats}},
\code{\link{theo_md}},
\code{\link{theo_sd}},
\code{\link{warfarin}}
}
\concept{nlmixr datasets}
\keyword{datasets}
