% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foceiFit.R
\name{yeoJohnson}
\alias{yeoJohnson}
\title{Yeo-Johnson Transformation}
\usage{
yeoJohnson(x, lambda = 1)
}
\arguments{
\item{x}{data to transform}

\item{lambda}{Cox-box lambda parameter}
}
\value{
Yeo-Johnson  Transformed Data
}
\description{
Yeo-Johnson Transformation
}
\author{
Matthew L. Fidler
}
