% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{addCwres}
\alias{addCwres}
\title{Add CWRES}
\usage{
addCwres(fit, updateObject = TRUE)
}
\arguments{
\item{fit}{nlmixr fit without WRES/CWRES}

\item{updateObject}{Boolean indicating if the original fit object should be updated. By default this is true.}
}
\value{
fit with CWRES
}
\description{
This returns a new fit object with CWRES attached
}
\author{
Matthew L. Fidler
}
