% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofv.R
\name{addCwres}
\alias{addCwres}
\title{Add CWRES}
\usage{
addCwres(fit, updateObject = TRUE, envir = globalenv())
}
\arguments{
\item{fit}{nlmixr fit without WRES/CWRES}

\item{updateObject}{Boolean indicating if the original fit object
should be updated. By default this is true.}

\item{envir}{Environment that should be checked for object to
update.  By default this is the global environment.}
}
\value{
fit with CWRES
}
\description{
This returns a new fit object with CWRES attached
}
\author{
Matthew L. Fidler
}
