% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npde.R
\name{addNpde}
\alias{addNpde}
\alias{.npdeCalc}
\alias{.npdeSim}
\title{NPDE calculation for nlmixr}
\usage{
addNpde(
  object,
  nsim = 1000,
  ties = TRUE,
  seed = 1009,
  updateObject = TRUE,
  cholSEtol = (.Machine$double.eps)^(1/3),
  ...
)

.npdeCalc(
  object,
  nsim = 1000,
  ties = TRUE,
  seed = 1009,
  updateObject = TRUE,
  cholSEtol = (.Machine$double.eps)^(1/3),
  ...,
  sim
)

.npdeSim(
  object,
  nsim = 1000,
  ties = TRUE,
  seed = 1009,
  updateObject = TRUE,
  cholSEtol = (.Machine$double.eps)^(1/3),
  ...
)
}
\arguments{
\item{object}{nlmixr fit object}

\item{nsim}{Number of simulations.  By default this is 300}

\item{ties}{When TRUE, the npde distribution can have ties.  When
FALSE, the npde distribution uses uniform random numbers to
prevent ties.}

\item{seed}{Seed for running nlmixr simulation.  By default 1009}

\item{updateObject}{Boolean indicating if original object should be updated.  By default this is TRUE.}

\item{cholSEtol}{tolerance for Generalized Cholesky
Decomposition.  Defaults to suggested (.Machine$double.eps)^(1/3)}

\item{...}{Other ignored parameters.}

\item{sim}{With `.ndpeCalc` the RxODE simulated values to
calculate `npde`}
}
\value{
New nlmixr fit object
}
\description{
NPDE calculation for nlmixr
}
\details{
`addNpde` is the user-facing function to add the NPDE

`.npdeSim` is the internal function that simulates everything for
the npde calculation

`.npdeCalc` calculates the extra npde items based in the `sim` parameter
}
\examples{

\donttest{

one.cmt <- function() {
  ini({
    ## You may label each parameter with a comment
    tka <- 0.45 # Log Ka
    tcl <- log(c(0, 2.7, 100)) # Log Cl
    ## This works with interactive models
    ## You may also label the preceding line with label("label text")
    tv <- 3.45; label("log V")
    ## the label("Label name") works with all models
    eta.ka ~ 0.6
    eta.cl ~ 0.3
    eta.v ~ 0.1
    add.sd <- 0.7
  })
  model({
    ka <- exp(tka + eta.ka)
    cl <- exp(tcl + eta.cl)
    v <- exp(tv + eta.v)
    linCmt() ~ add(add.sd)
  })
}

f <- nlmixr(one.cmt, theo_sd, "saem")

# even though you may have forgotten to add the NPDE, you can add it to the data.frame:

f <- addNpde(f)

}
}
\author{
Matthew L. Fidler
}
