% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foceiFit.R, R/nlme_fit.R
\name{as.focei}
\alias{as.focei}
\alias{as.focei.nlmixrNlme}
\title{Convert fit to FOCEi style fit}
\usage{
as.focei(object, uif, pt = proc.time(), ..., data, calcResid = TRUE)

\method{as.focei}{nlmixrNlme}(
  object,
  uif,
  pt = proc.time(),
  ...,
  data,
  calcResid = TRUE,
  nobs2 = 0,
  keep = NULL,
  drop = NULL,
  table = tableControl()
)
}
\arguments{
\item{object}{Fit object to convert to FOCEi-style fit.}

\item{uif}{Unified Interface Function}

\item{pt}{Proc time object}

\item{...}{Other Parameters}

\item{data}{The data to pass to the FOCEi translation.}

\item{calcResid}{A boolean to indicate if the CWRES residuals
should be calculated}

\item{nobs2}{Number of observations without EVID=2}

\item{keep}{Columns to keep from either the input dataset. For the
input dataset, if any records are added to the data LOCF (Last
Observation Carried forward) imputation is performed.}

\item{drop}{Columns to drop from the output}
}
\value{
A FOCEi fit style object.
}
\description{
Convert fit to FOCEi style fit
}
\author{
Matthew L. Fidler
}
