% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{as.nlme}
\alias{as.nlme}
\title{Return composite nlme/focei to nlme}
\usage{
as.nlme(x)
}
\arguments{
\item{x}{nlme/focei object from common UI.}
}
\value{
nlme object or NULL
}
\description{
Return composite nlme/focei to nlme
}
\author{
Matthew L. Fidler
}
\keyword{internal}

