% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{as.saem}
\alias{as.saem}
\title{Return composite saem/focei to saem}
\usage{
as.saem(x)
}
\arguments{
\item{x}{saem/focei object from common UI.}
}
\value{
saem object or NULL
}
\description{
Return composite saem/focei to saem
}
\author{
Matthew L. Fidler
}
\keyword{internal}
