% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bootplot}
\alias{bootplot}
\title{Produce trace-plot for fit if applicable}
\usage{
bootplot(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{other parameters}
}
\value{
Fit traceplot or nothing.
}
\description{
Produce trace-plot for fit if applicable
}
\author{
Vipul Mann,  Matthew L. Fidler
}
