% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapFit}
\alias{bootstrapFit}
\title{Bootstrap nlmixr fit}
\usage{
bootstrapFit(
  fit,
  nboot = 500,
  nSampIndiv,
  stratVar,
  stdErrType = c("perc", "se"),
  ci = 0.95,
  pvalues = NULL,
  restart = FALSE,
  plotHist = TRUE
)
}
\arguments{
\item{fit}{the nlmixr fit object}

\item{nboot}{an integer giving the number of bootstrapped models to be fit; default value is 100}

\item{nSampIndiv}{an integer specifying the number of samples in each bootstrapped sample; default is the number of unique subjects in the original dataset}

\item{pvalues}{a vector of pvalues indicating the probability of each subject to get selected; default value is NULL implying that probability of each subject is the same}

\item{restart}{a boolean that indicates if a previous session has to be restarted; default value is FALSE}
}
\description{
Bootstrap input dataset and rerun the model to get confidence bounds and aggregated parameters
}
\examples{
\dontrun{
one.cmt <- function() {
  ini({
    ## You may label each parameter with a comment
    tka <- 0.45 # Log Ka
    tcl <- 1 # Log Cl
    ## This works with interactive models
    ## You may also label the preceding line with label("label text")
    tv <- 3.45
    label("log V")
    ## the label("Label name") works with all models
    eta.ka ~ 0.6
    eta.cl ~ 0.3
    eta.v ~ 0.1
    add.sd <- 0.7
  })
  model({
    ka <- exp(tka + eta.ka)
    cl <- exp(tcl + eta.cl)
    v <- exp(tv + eta.v)
    linCmt() ~ add(add.sd)
  })
}

fit <- nlmixr(one.cmt, theo_sd, "focei")

bootstrapFit(fit)
bootstrapFit(fit, nboot = 5, restart = TRUE) # overwrites any of the existing data or model files
bootstrapFit(fit, nboot = 7) # resumes fitting using the stored data and model files
}
}
\author{
Vipul Mann, Matthew Fidler
}
