% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnlmm.R
\name{calcCov}
\alias{calcCov}
\title{Calculate gnlmm variance-covariance matrix of fixed effects}
\usage{
calcCov(fit, method = 1, trace = FALSE)
}
\arguments{
\item{fit}{a gnlmm fit object}

\item{method}{method for calculating variance-covariance matrix}

\item{trace}{logical whether to trace the iterations}
}
\value{
variance-covariance matrix of model parameters
}
\description{
Calculate variance-covariance matrix of fixed effects after a gnlmm() fit
}
\examples{
\dontrun{
calcCov(fit)
}
}

