% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{fitted.focei.fit}
\alias{fitted.focei.fit}
\title{Extract the fitted values from the model}
\usage{
\method{fitted}{focei.fit}(object, ..., population = FALSE,
  type = c("fitted", "Vi", "Vfo"))
}
\arguments{
\item{object}{Fit object}

\item{...}{other parameters}

\item{population}{when true (default false), calculate the
    population predictions; When false, calculate the individual
    predictions.

    If this is a matrix with the number columns equal to the
    number of ETAs in model, and the number of rows equals to the
    number of subjects in the dataset, then these etas will be
    used to fit the data.}
}
\value{
Individual predictions
}
\description{
Extract the fitted values from the model
}
\author{
Matthew L. Fidler
}
