% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R, R/nlme_fit.R
\name{focei.eta}
\alias{focei.eta}
\alias{focei.eta.nlmixrNlme}
\title{Get the FOCEi theta or eta specification for model.}
\usage{
focei.eta(object, uif, ...)

\method{focei.eta}{nlmixrNlme}(object, ...)
}
\arguments{
\item{object}{Fit object}

\item{uif}{User interface function or object}

\item{...}{Other parameters}
}
\value{
List for the OMGA list in FOCEi
}
\description{
Get the FOCEi theta or eta specification for model.
}
\author{
Matthew L. Fidler
}
