% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R, R/utils_cmt_ode12.R
\name{focei.fit}
\alias{focei.fit}
\alias{focei.fit}
\title{FOCEI Fit for nlmixr}
\usage{
focei.fit(data, inits, PKpars, diag.xform = c("sqrt", "log", "identity"),
  optim = c("newuoa", "lbfgsb3", "nlminb", "nelder-mead"), model = list(),
  control = list())

focei.fit(data, inits, PKpars, diag.xform = c("sqrt", "log", "identity"),
  optim = c("newuoa", "lbfgsb3", "nlminb", "nelder-mead"), model = list(),
  control = list())
}
\arguments{
\item{data}{Data to fit}

\item{inits}{Initilization list}

\item{PKpars}{Pk Parameters}

\item{diag.xform}{The form of the diagonal of the Omega matrix to
be estimated.  This can be "sqrt" "log" or "identity".}

\item{optim}{The optimization method to be used}

\item{model}{The RxODE model to use}

\item{control}{Control list}

\item{pred}{The Prediction function}

\item{err}{The Error function}

\item{lower}{Lower bounds}

\item{upper}{Upper Bounds}

\item{calculate.vars}{is a list of variables that will be
calculated after the FOCEI estimation is complete.}

\item{llik}{log-likelihood function}

\item{data}{data to be fitted}

\item{inits}{initial values}

\item{system}{an optional (compiled) RxODE object}

\item{syspar}{function: calculation of PK parameters}

\item{diag.xform}{transformation to diagonal elements of OMEGA during fitting}

\item{...}{additional options}

\item{control}{additional optimization options}
}
\value{
A focei fit object
}
\description{
FOCEI Fit for nlmixr

Fit a generalized nonlinear mixed-effect model by adapative Gaussian quadrature (AQD)
}
\details{
Fit a generalized nonlinear mixed-effect model by adapative Gaussian quadrature (AGQ)
}
\examples{
\dontrun{
llik <- function()
{
	if (group==1) lp = THETA[1]+THETA[2]*logtstd+ETA[1]
	else          lp = THETA[3]+THETA[4]*logtstd+ETA[1]
	lam = exp(lp)
	dpois(y, lam, log=TRUE)
}
inits = list(THTA=c(1,1,1,1), OMGA=list(ETA[1]~1))

fit = gnlmm(llik, pump, inits, 
	control=list(
	    reltol.outer=1e-4,
		optim.outer="nmsimplex",
		nAQD=5
	)
)



llik <- function()
{
	lp = THETA[1]*x1+THETA[2]*x2+(x1+x2*THETA[3])*ETA[1]
	p = pnorm(lp)
	dbinom(x, m, p, log=TRUE)
}
inits = list(THTA=c(1,1,1), OMGA=list(ETA[1]~1))

gnlmm(llik, rats, inits, control=list(nAQD=7))


ode <- "
d/dt(depot) =-KA*depot;
d/dt(centr) = KA*depot - KE*centr;
"
sys1 = RxODE(ode)

pars <- function()
{
	CL = exp(THETA[1] + ETA[1])#; if (CL>100) CL=100
	KA = exp(THETA[2] + ETA[2])#; if (KA>20) KA=20
	KE = exp(THETA[3])
	V  = CL/KE
	sig2 = exp(THETA[4])
}
llik <- function() {
	pred = centr/V
	dnorm(DV, pred, sd=sqrt(sig2), log=TRUE)
}
inits = list(THTA=c(-3.22, 0.47, -2.45, 0))
inits$OMGA=list(ETA[1]~.027, ETA[2]~.37)
#inits$OMGA=list(ETA[1]+ETA[2]~c(.027, .01, .37))
theo <- read.table("theo_md.txt", head=TRUE)

fit = gnlmm(llik, theo, inits, pars, sys1, 
	control=list(trace=TRUE, nAQD=5))

cv = calcCov(fit)
cbind(fit$par[fit$nsplt==1], sqrt(diag(cv)))

}
}
\author{
Matthew L. Fidler and Wenping Wang

Wenping Wang
}
