% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{focei.fit}
\alias{focei.fit}
\title{FOCEI Fit for nlmixr}
\usage{
focei.fit(data, inits, PKpars, diag.xform = c("sqrt", "log", "identity"),
  optim = c("bobyqa", "nlminb"), model = NULL, pred = NULL, err = NULL,
  lower = -Inf, upper = Inf, control = list(),
  calculate.vars = c("pred", "ipred", "ires", "res", "iwres", "wres", "cwres",
  "cpred", "cres"))
}
\arguments{
\item{data}{Data to fit}

\item{inits}{Initilization list}

\item{PKpars}{Pk Parameters}

\item{diag.xform}{The form of the diagonal of the Omega matrix to
be estimated.  This can be "sqrt" "log" or "identity".}

\item{optim}{The optimization method to be used}

\item{model}{The RxODE model to use}

\item{pred}{The Prediction function}

\item{err}{The Error function}

\item{lower}{Lower bounds}

\item{upper}{Upper Bounds}

\item{control}{Control list}

\item{calculate.vars}{is a list of variables that will be
calculated after the FOCEI estimation is complete.}
}
\value{
A focei fit object
}
\description{
FOCEI Fit for nlmixr
}
\author{
Matthew L. Fidler and Wenping Wang
}
