% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{focei.fit}
\alias{focei.fit}
\alias{focei.fit.nlmixr.ui.nlme}
\alias{focei.fit.nlmixr.ui.focei.fit}
\alias{focei.fit.nlmixrUI}
\alias{focei.fit.function}
\alias{focei.fit.data.frame}
\title{FOCEI Fit for nlmixr}
\usage{
focei.fit(data, inits, PKpars, diag.xform = c("sqrt", "log", "identity"),
  optim = c("lbfgs", "L-BFGS-B", "bobyqa", "lbfgsb3", "nlminb"),
  model = NULL, pred = NULL, err = NULL, lower = -Inf, upper = Inf,
  control = list(), calculate.vars = c("pred", "ipred", "ires", "res",
  "iwres", "wres", "cwres", "cpred", "cres"), theta.names = NULL,
  eta.names = NULL, ...)

\method{focei.fit}{nlmixr.ui.nlme}(data, inits, ...)

\method{focei.fit}{nlmixr.ui.focei.fit}(data, inits, ...)

\method{focei.fit}{nlmixrUI}(data, inits, ...)

\method{focei.fit}{function}(data, inits, ...)

\method{focei.fit}{data.frame}(...)
}
\arguments{
\item{data}{Data to fit}

\item{inits}{Initialization list}

\item{PKpars}{Pk Parameters}

\item{diag.xform}{The form of the diagonal of the Omega matrix to
be estimated.  This can be "sqrt" "log" or "identity".}

\item{optim}{The optimization method to be used}

\item{model}{The RxODE model to use}

\item{pred}{The Prediction function}

\item{err}{The Error function}

\item{lower}{Lower bounds}

\item{upper}{Upper Bounds}

\item{control}{Control list}

\item{calculate.vars}{is a list of variables that will be
calculated after the FOCEI estimation is complete.}

\item{theta.names}{Names of the thetas to be used in the final object}

\item{eta.names}{Eta names to be used in the final object}

\item{...}{Ignored parameters}
}
\value{
A focei fit object
}
\description{
FOCEI Fit for nlmixr
}
\author{
Matthew L. Fidler and Wenping Wang
}
