% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.quad.R
\name{gauss.quad}
\alias{gauss.quad}
\title{Sets nodes and weigts of Gauss-Hermite quadrature}
\usage{
gauss.quad(n)
}
\arguments{
\item{n}{numbr of nodes}
}
\value{
a list of nodes and weigts of Gauss-Hermite quadrature
}
\description{
Sets nodes and weigts of Gauss-Hermite quadrature
}
\examples{
\dontrun{
gauss.quad(5)
}
}
