% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynmodel.R
\name{gof}
\alias{gof}
\alias{plot.dyn.ID}
\title{Plot of a non-population dynamic model fit}
\usage{
gof(x, ...)

\method{plot}{dyn.ID}(x, ...)
}
\arguments{
\item{x}{a dynamodel fit object}

\item{...}{additional arguments}
}
\value{
nothing, displays a goodness of fit plot for dynmodely
}
\description{
Plot of a non-population dynamic model fit
}
