% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{model}
\alias{model}
\title{nlmixr model block}
\usage{
model(model, ...)
}
\arguments{
\item{model}{Model specification}
}
\value{
Parsed UI object
}
\description{
nlmixr model block
}
\author{
Matthew L. Fidler
}
