% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n1qn1}
\alias{n1qn1}
\title{n1qn1 optimization}
\usage{
n1qn1(call_eval, call_grad, vars, environment = parent.frame(1), ...,
  epsilon = .Machine$double.eps, max_iterations = 100, nsim = 100,
  imp = 0, invisible = NULL, zm = NULL, restart = FALSE,
  assign = FALSE, print.functions = FALSE)
}
\arguments{
\item{call_eval}{Objective function}

\item{call_grad}{Gradient Function}

\item{vars}{Initial starting point for line search}

\item{environment}{Environment where call_eval/call_grad are evaluated.}

\item{max_iterations}{Number of iterations}

\item{nsim}{Number of function evaluations}

\item{imp}{Verbosity of messages.}

\item{zm}{Prior Hessian (in compressed format)}

\item{restart}{Is this an estimation restart?}

\item{assign}{Assign hessian to c.hess in environment environment? (Default FALSE)}
}
\description{
n1qn1 optimization
}
\author{
C. Lemarechal, Wenping Wang & Matthew L. Fidler
}
