% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nlmixr.fit}
\alias{nlmixr.fit}
\title{Fit a nlmixr model}
\usage{
\method{nlmixr}{fit}(uif, data, est = "nlme", ..., sum.prod = FALSE,
  focei.translate = TRUE)
}
\arguments{
\item{uif}{Parsed nlmixr model (by \code{nlmixr(mod.fn)}).}

\item{data}{Dataset to estimate}

\item{est}{Estimation method}

\item{...}{Parameters passed to estimation method.}

\item{focei.translate}{Translate the model to FOCEi and then run
the tables and objective function so that different estimation
methodologies are comparable through OBJF.}
}
\value{
nlmixr fit object
}
\description{
Fit a nlmixr model
}
\author{
Matthew L. Fidler
}
