% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nlmixr}
\alias{nlmixr}
\alias{nlmixr.function}
\alias{nlmixr.nlmixrUI}
\alias{nlmixr.nlmixr.ui.nlme}
\alias{nlmixr.nlmixr.ui.focei.fit}
\alias{nlmixr.nlmixr.ui.saem}
\title{nlmixr fits population PK and PKPD non-linear mixed effects models.}
\usage{
nlmixr(object, data, est = "nlme", control = list(), ...)

\method{nlmixr}{function}(object, data, est = "nlme", control = list(), ...)

\method{nlmixr}{nlmixrUI}(object, data, est = "nlme", control = list(), ...)

\method{nlmixr}{nlmixr.ui.nlme}(object, data, est = "nlme", ...)

\method{nlmixr}{nlmixr.ui.focei.fit}(object, data, est = "nlme", ...)

\method{nlmixr}{nlmixr.ui.saem}(object, data, est = "nlme", ...)
}
\arguments{
\item{object}{Fitted object or function specifying the model.}

\item{data}{Dataset to estimate.  Needs to be RxODE compatible in
EVIDs.}

\item{est}{Estimation method}

\item{control}{Estimation control options.  They could be
\code{\link[nlme]{nlmeControl}}, \code{\link{saemControl}}}

\item{...}{Other parameters}
}
\value{
Either a nlmixr model or a nlmixr fit object
}
\description{
nlmixr is an R package for fitting population pharmacokinetic (PK)
and pharmacokinetic-pharmacodynamic (PKPD) models.
}
\details{
The nlmixr generalized function allows common access to the nlmixr
estimation routines.
}
\author{
Matthew L. Fidler
}
