% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nlmixr}
\alias{nlmixr}
\alias{nlmixr.function}
\alias{nlmixr.nlmixrUI}
\alias{nlmixr.nlmixr.ui.nlme}
\alias{nlmixr.nlmixr.ui.focei.fit}
\title{nlmixr generalized function}
\usage{
nlmixr(object, data, est = "nlme", ...)

\method{nlmixr}{function}(object, data, est = "nlme", ...)

\method{nlmixr}{nlmixrUI}(object, data, est = "nlme", ...)

\method{nlmixr}{nlmixr.ui.nlme}(object, data, est = "nlme", ...)

\method{nlmixr}{nlmixr.ui.focei.fit}(object, data, est = "nlme", ...)
}
\arguments{
\item{object}{Fitted object or}

\item{data}{Data for fit.}

\item{est}{Estimation routine.}
}
\value{
Either a nlmixr model or a nlmixr fit object
}
\description{
nlmixr generalized function
}
\author{
Matthew L. Fidler
}
