% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{nlmixrBounds}
\alias{nlmixrBounds}
\title{Extract the Nlmixr bound information from a function.}
\usage{
nlmixrBounds(fun)
}
\arguments{
\item{fun}{Function to extract bound information from.}
}
\value{
a dataframe with bound information.
}
\description{
Extract the Nlmixr bound information from a function.
}
\seealso{
Other nlmixrBounds: 
\code{\link{nlmixrBoundsParser}()}
}
\author{
Matthew L. Fidler
}
\concept{nlmixrBounds}
\keyword{internal}
