% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{nlmixrBoundsParser}
\alias{nlmixrBoundsParser}
\alias{nlmixrBoundsParser.function}
\alias{nlmixrBoundsParser.{}
\alias{nlmixrBoundsParser.(}
\alias{nlmixrBoundsParser.<-}
\alias{nlmixrBoundsParser.=}
\alias{nlmixrBoundsParser.call}
\alias{nlmixrBoundsParser.numeric}
\alias{nlmixrBoundsParser.integer}
\title{Functions to assist with setting initial conditions and boundaries}
\usage{
nlmixrBoundsParser(x)

\method{nlmixrBoundsParser}{`function`}(x)

\method{nlmixrBoundsParser}{`{`}(x)

\method{nlmixrBoundsParser}{`(`}(x)

\method{nlmixrBoundsParser}{`<-`}(x)

\method{nlmixrBoundsParser}{`=`}(x)

\method{nlmixrBoundsParser}{call}(x)

\method{nlmixrBoundsParser}{numeric}(x)

\method{nlmixrBoundsParser}{integer}(x)
}
\arguments{
\item{x}{the object to attempt extraction from}
}
\value{
A list with how the object will be used
}
\description{
These functions are not intended to be called by a user.  They are intended
to be internal to nlmixr.
}
\section{Methods (by class)}{
\itemize{
\item \code{function}: For functions, apply to the function body

\item \code{{}: For function bodies and similar.

\item \code{(}: For function bodies and similar.

\item \code{<-}: Assignments to thetas with names

\item \code{=}: Assignments to thetas with names

\item \code{call}: Assignments to thetas without names,
assignment to omegas with or without names, or setting of attributes.

\item \code{numeric}: Assignments of numbers to thetas without names

\item \code{integer}: Assignments of numbers to thetas without names
}}

\seealso{
Other nlmixrBounds: 
\code{\link{nlmixrBounds}()}
}
\concept{nlmixrBounds}
\keyword{internal}
