% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{nlmixrBoundsParserTheta}
\alias{nlmixrBoundsParserTheta}
\title{Forms of fixed that are allowed are:}
\usage{
nlmixrBoundsParserTheta(x, currentData)
}
\description{
* `fixed(1)`
* `fixed(1, 2)`
* `fixed(1, 2, 3)`
* `c(1, fixed)`
* `c(1, 2, fixed)`
* `c(1, 2, 3, fixed)`
* `c(1, fixed(2))`
* `c(1, fixed(2), 3)`
}
\details{
Where "fixed" can be "FIX", "FIXED", "fix", or "fixed".
}
