% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nlmixrUI.nlmefun}
\alias{nlmixrUI.nlmefun}
\title{Create the nlme parameter transform function from the UI object.}
\usage{
nlmixrUI.nlmefun(object)
}
\arguments{
\item{object}{UI object}
}
\value{
parameter function for nlme
}
\description{
Create the nlme parameter transform function from the UI object.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
