% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{nmDataConvert}
\alias{nmDataConvert}
\title{Convert data to RxODE format (depreciated)}
\usage{
nmDataConvert(data)
}
\arguments{
\item{data}{Data to "convert"}
}
\value{
Exact same data as was inputRxODE::
}
\description{
Convert data to RxODE format (depreciated)
}
\keyword{internal}
