% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.dyn.ID}
\alias{plot.dyn.ID}
\title{Plot of a non-population dynamic model fit}
\usage{
\method{plot}{dyn.ID}(x, ...)
}
\arguments{
\item{x}{a dynamodel fit object}

\item{...}{additional arguments}
}
\description{
Plot of a non-population dynamic model fit
}
