% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{plot.focei.fit}
\alias{plot.focei.fit}
\title{Plot a focei.fit plot}
\usage{
\method{plot}{focei.fit}(x, ...)
}
\arguments{
\item{x}{a focei fit object}

\item{...}{additional arguments}
}
\description{
Plot some standard goodness of fit plots for the focei fitted object
}
\author{
Wenping Wang & Matthew Fidler
}

