% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.nlmixrFitData}
\alias{plot.nlmixrFitData}
\title{Plot a nlmixr data object}
\usage{
\method{plot}{nlmixrFitData}(x, ...)
}
\arguments{
\item{x}{a focei fit object}

\item{...}{additional arguments}
}
\description{
Plot some standard goodness of fit plots for the focei fitted object
}
\author{
Wenping Wang & Matthew Fidler
}
