% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem_fit.R
\name{plot.saemFit}
\alias{plot.saemFit}
\title{plot a saem fit}
\usage{
\method{plot}{saemFit}(x, ...)
}
\arguments{
\item{x}{a saemFit object}

\item{...}{others}
}
\value{
a list
}
\description{
plot a saem fit
}

