% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precondition.R
\name{preconditionFit}
\alias{preconditionFit}
\title{Linearly re-parameterize the model to be less sensitive to rounding errors}
\usage{
preconditionFit(fit, estType = c("full", "posthoc", "none"))
}
\arguments{
\item{fit}{A nlmixr fit to be preconditioned}

\item{estType}{Once the fit has been linearly reparametrized,
should a "full" estimation, "posthoc" estimation or simply a
estimation of the covariance matrix "none" before the fit is updated}
}
\description{
Linearly re-parameterize the model to be less sensitive to rounding errors
}
