% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dyn.mcmc}
\alias{print.dyn.mcmc}
\title{summary of a dynmodel fit}
\usage{
\method{print}{dyn.mcmc}(x, ...)
}
\arguments{
\item{x}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
summary of a dynmodel fit
}

