% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynmodel.R
\name{summary.dyn.mcmc}
\alias{summary.dyn.mcmc}
\alias{print.dyn.mcmc}
\title{Print summary of a non-population dynamic model fit using mcmc}
\usage{
\method{summary}{dyn.mcmc}(object, ...)

\method{print}{dyn.mcmc}(x, ...)
}
\arguments{
\item{...}{additional arguments}

\item{x, object}{a dynmodel fit object}
}
\value{
invisibly return original object
}
\description{
Print summary of a non-population dynamic model fit using mcmc
}
