% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cmt_ode4.R
\name{print.focei.fit}
\alias{print.focei.fit}
\title{Print a gnlmm fit}
\usage{
\method{print}{focei.fit}(fit)
}
\arguments{
\item{x}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
Print a generalized non-linear mixed effect model fit
}
