% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{print.focei.fit}
\alias{print.focei.fit}
\title{Print a focei fit}
\usage{
\method{print}{focei.fit}(x, ...)
}
\arguments{
\item{x}{a focei model fit}

\item{...}{additional arguments}
}
\description{
Print a first-order conditional non-linear mixed effect model with
interaction fit
}
