% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/simulate.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{RxODE}
\alias{reexports}
\alias{rxSolve}
\alias{reexports}
\alias{nlme}
\alias{reexports}
\alias{ACF}
\alias{reexports}
\alias{VarCorr}
\alias{reexports}
\alias{augPred}
\alias{reexports}
\alias{fixef}
\alias{reexports}
\alias{fixed.effects}
\alias{reexports}
\alias{ranef}
\alias{reexports}
\alias{random.effects}
\alias{reexports}
\alias{varWeights}
\alias{reexports}
\alias{nlmeControl}
\alias{reexports}
\alias{eventTable}
\alias{reexports}
\alias{add.dosing}
\alias{reexports}
\alias{add.sampling}
\alias{reexports}
\alias{rxSolve}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{nlme}{\code{\link[nlme]{nlme}}, \code{\link[nlme]{ACF}}, \code{\link[nlme]{VarCorr}}, \code{\link[nlme]{augPred}}, \code{\link[nlme]{fixef}}, \code{\link[nlme]{fixed.effects}}, \code{\link[nlme]{ranef}}, \code{\link[nlme]{random.effects}}, \code{\link[nlme]{varWeights}}, \code{\link[nlme]{nlmeControl}}}

  \item{RxODE}{\code{\link[RxODE]{RxODE}}, \code{\link[RxODE]{rxSolve}}, \code{\link[RxODE]{eventTable}}, \code{\link[RxODE]{add.dosing}}, \code{\link[RxODE]{add.sampling}}, \code{\link[RxODE]{rxSolve}}}
}}

