% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/simulate.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{RxODE}
\alias{logit}
\alias{expit}
\alias{probit}
\alias{probitInv}
\alias{rxSolve}
\alias{rxClean}
\alias{rxCat}
\alias{nlme}
\alias{ACF}
\alias{VarCorr}
\alias{getVarCov}
\alias{augPred}
\alias{fixef}
\alias{fixed.effects}
\alias{ranef}
\alias{random.effects}
\alias{nlmeControl}
\alias{eventTable}
\alias{add.dosing}
\alias{add.sampling}
\alias{rxPermissive}
\alias{pdDiag}
\alias{pdSymm}
\alias{pdLogChol}
\alias{pdIdent}
\alias{pdCompSymm}
\alias{pdBlocked}
\alias{pdNatural}
\alias{pdConstruct}
\alias{pdFactor}
\alias{pdMat}
\alias{pdMatrix}
\alias{reStruct}
\alias{varWeights}
\alias{varPower}
\alias{varFixed}
\alias{varFunc}
\alias{varExp}
\alias{varConstPower}
\alias{varIdent}
\alias{varComb}
\alias{groupedData}
\alias{getData}
\alias{rxSymPySetupPred}
\alias{et}
\alias{rxParams}
\alias{rxParam}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{nlme}{\code{\link[nlme]{ACF}}, \code{\link[nlme]{VarCorr}}, \code{\link[nlme]{augPred}}, \code{\link[nlme]{fixed.effects}}, \code{\link[nlme:fixed.effects]{fixef}}, \code{\link[nlme]{getData}}, \code{\link[nlme]{getVarCov}}, \code{\link[nlme]{groupedData}}, \code{\link[nlme]{nlme}}, \code{\link[nlme]{nlmeControl}}, \code{\link[nlme]{pdBlocked}}, \code{\link[nlme]{pdCompSymm}}, \code{\link[nlme]{pdConstruct}}, \code{\link[nlme]{pdDiag}}, \code{\link[nlme]{pdFactor}}, \code{\link[nlme]{pdIdent}}, \code{\link[nlme]{pdLogChol}}, \code{\link[nlme]{pdMat}}, \code{\link[nlme]{pdMatrix}}, \code{\link[nlme]{pdNatural}}, \code{\link[nlme]{pdSymm}}, \code{\link[nlme]{random.effects}}, \code{\link[nlme:random.effects]{ranef}}, \code{\link[nlme]{reStruct}}, \code{\link[nlme]{varComb}}, \code{\link[nlme]{varConstPower}}, \code{\link[nlme]{varExp}}, \code{\link[nlme]{varFixed}}, \code{\link[nlme]{varFunc}}, \code{\link[nlme]{varIdent}}, \code{\link[nlme]{varPower}}, \code{\link[nlme]{varWeights}}}

  \item{RxODE}{\code{\link[RxODE]{RxODE}}, \code{\link[RxODE]{add.dosing}}, \code{\link[RxODE]{add.sampling}}, \code{\link[RxODE]{et}}, \code{\link[RxODE]{et}}, \code{\link[RxODE]{eventTable}}, \code{\link[RxODE:logit]{expit}}, \code{\link[RxODE]{logit}}, \code{\link[RxODE]{probit}}, \code{\link[RxODE:probit]{probitInv}}, \code{\link[RxODE]{rxCat}}, \code{\link[RxODE]{rxClean}}, \code{\link[RxODE:rxParams]{rxParam}}, \code{\link[RxODE]{rxParams}}, \code{\link[RxODE]{rxPermissive}}, \code{\link[RxODE]{rxSolve}}, \code{\link[RxODE]{rxSolve}}, \code{\link[RxODE:rxSEinner]{rxSymPySetupPred}}}
}}

