% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{residuals.focei.fit}
\alias{residuals.focei.fit}
\title{Extract residuals from the FOCEI fit}
\usage{
\method{residuals}{focei.fit}(object, ..., type = c("ires", "res", "iwres",
  "wres", "cwres"))
}
\arguments{
\item{object}{focei.fit object}

\item{...}{Additional arguments}

\item{type}{Type of resduals fitted.}
}
\value{
residuals
}
\description{
Extract residuals from the FOCEI fit
}
\author{
Matthew L. Fidler
}
