% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foceiFit.R
\name{setOfv}
\alias{setOfv}
\title{Set Objective function type for a nlmixr object}
\usage{
setOfv(x, type)
}
\arguments{
\item{x}{nlmixr fit object}

\item{type}{Type of objective function to use for AIC, BIC, and $objective}
}
\value{
Nothing
}
\description{
Set Objective function type for a nlmixr object
}
\author{
Matthew L. Fidler
}
