% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{simulate.focei.fit}
\alias{simulate.focei.fit}
\title{Simulate response based on FOCEi model's datathe fitted object's dataset}
\usage{
\method{simulate}{focei.fit}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{Focei object}

\item{nsim}{Number of simulated parameters}

\item{seed}{Seed to start with (if specified)}

\item{...}{Other parameters}
}
\value{
New dataset based on original dataset
}
\description{
Simulate response based on FOCEi model's datathe fitted object's dataset
}
\author{
Matthew L. Fidler
}
