% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.dyn.ID}
\alias{summary.dyn.ID}
\title{summary of a dynmodel fit}
\usage{
\method{summary}{dyn.ID}(object, ...)
}
\arguments{
\item{object}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
summary of a dynmodel fit
}

