% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.dyn.mcmc}
\alias{summary.dyn.mcmc}
\title{Summary of a non-population dynamic model fit using mcmc}
\usage{
\method{summary}{dyn.mcmc}(object, ...)
}
\arguments{
\item{object}{a dynmodel fit object}

\item{...}{additional arguments}
}
\description{
Summary of a non-population dynamic model fit using mcmc
}

