% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem_fit.R
\name{summary.saemFit}
\alias{summary.saemFit}
\title{Print an SAEM model fit summary}
\usage{
\method{summary}{saemFit}(object, ...)
}
\arguments{
\item{object}{a saemFit object}

\item{...}{others}
}
\value{
a list
}
\description{
Print an SAEM model fit summary
}
