% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_theo_sd.R
\docType{data}
\name{theo_sd}
\alias{theo_sd}
\title{Multiple dose theophylline PK data}
\format{
A data frame with 144 rows by 7 columns

\describe{
  \item{ID}{Subject ID}
  \item{TIME}{Time (hrs)}
  \item{DV}{Dependent Variable, theophylline concentration}
  \item{AMT}{Dose Amount/kg}
  \item{EVID}{RxODE/nlmixr event ID (not NONMEM's)}
  \item{CMT}{Compartment Number}
  \item{WT}{Weight (kg)}
}
}
\source{
NONMEM/nlme
}
\usage{
theo_sd
}
\description{
This data set is  the day 1 concentrations of the
theophylline data that is included in the nlme/NONMEM.
}
\seealso{
Other nlmixr datasets: 
\code{\link{Bolus_1CPTMM}},
\code{\link{Bolus_1CPT}},
\code{\link{Bolus_2CPTMM}},
\code{\link{Bolus_2CPT}},
\code{\link{Infusion_1CPT}},
\code{\link{Oral_1CPT}},
\code{\link{Wang2007}},
\code{\link{pheno_sd}},
\code{\link{rats}},
\code{\link{theo_md}},
\code{\link{warfarin}}
}
\concept{nlmixr datasets}
\keyword{datasets}
