% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{traceplot}
\alias{traceplot}
\alias{traceplot.focei.fit}
\title{Produce trace-plot for fit if applicable}
\usage{
traceplot(x, ...)

\method{traceplot}{focei.fit}(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{other parameters}
}
\value{
Fit traceplot or nothing.
}
\description{
Produce trace-plot for fit if applicable
}
\author{
Rik Schoemaker, Wenping Wang & Matthew L. Fidler
}
