% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei_fit.R
\name{vcov.focei.fit}
\alias{vcov.focei.fit}
\title{Extract variance/covariance matrix for FOCEI fit}
\usage{
\method{vcov}{focei.fit}(object, ...)
}
\arguments{
\item{object}{Focei fit object}

\item{...}{ignored parameters}
}
\value{
variance/covariance matrix
}
\description{
Extract variance/covariance matrix for FOCEI fit
}
\author{
Matthew L. Fidler
}
