% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{vpc}
\alias{vpc}
\title{Visual predictive check (VPC)}
\usage{
vpc(fit, nsim = 100, condition = NULL)
}
\arguments{
\item{fit}{nlme fit object}

\item{nsim}{number of simulations}

\item{condition}{conditional variable}
}
\description{
Do visual predictive check (VPC) plots for nlme-base non-linear mixed effect models
}
\examples{
specs <- list(fixed=lKA+lCL+lV~1, random = pdDiag(lKA+lCL~1), start=c(lKA=0.5, lCL=-3.2, lV=-1))
fit <- nlme_lin_cmpt(theo_md, par_model=specs, ncmt=1, verbose=TRUE)
vpc(fit, nsim = 100, condition = NULL)
}
