% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{vpc}
\alias{vpc}
\alias{vpc.default}
\title{Vpc function for nlmixr}
\usage{
vpc(sim, ...)

\method{vpc}{default}(sim, ...)
}
\arguments{
\item{sim}{Observed data frame or fit object}

\item{...}{Other parameters}
}
\description{
Vpc function for nlmixr
}
