% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{vpc.ui}
\alias{vpc.ui}
\title{VPC based on ui model}
\usage{
vpc.ui(fit)
}
\arguments{
\item{fit}{nlmixr fit object}

\item{n}{Number of VPC simulations.  By default 100}

\item{...}{Args sent to rxSolve}
}
\description{
VPC based on ui model
}
\author{
Matthew L. Fidler
}
