% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wenping2007.R
\docType{data}
\name{Wang2007}
\alias{Wang2007}
\title{Simulated Data Set for comparing objective functions}
\format{A data frame with 20 rows and 3 rows

\describe{
  \item{ID}{Siumlated Subject ID}
  \item{Time}{Simulated Time}
  \item{Time}{Simulated Objective Function}
}}
\source{
Table 1 from Wang, Y \emph{Derivation of Various NONMEM estimation methods}. J Pharmacokinet Pharmacodyn (2007) 34:575-593.
}
\usage{
Wang2007
}
\description{
This is a simulated dataset from Wang2007 where various NONMEM
estimation methods (Lapclace FO, FOCE with and without
interaction) are described.
}
\keyword{datasets}
